% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinatize}
\alias{coordinatize}
\title{Convert a LAGOSUS flat file into a spatial object}
\usage{
coordinatize(
  dt,
  latname = "lake_lat_decdeg",
  longname = "lake_lon_decdeg",
  crs_in = 4326,
  crs_out = albers_conic()
)
}
\arguments{
\item{dt}{data.frame}

\item{latname}{character name of latitude column; default is "lake_lat_decdeg"}

\item{longname}{character name of longitude column; default is "lake_lon_decdeg"}

\item{crs_in}{numeric epsg code; default is 4326 WGS84-lat-long}

\item{crs_out}{numeric epsg code; default is albers_conic}
}
\description{
This function turns a data.frame into a spatial object using a reasonable
default projection.
}
\examples{
\dontrun{
dt <- lagosus_load("locus")
res <- coordinatize(dt$locus$locus_information)

library(maps)
map("state", xlim = c(-97.90363, -66.99892), ylim = c(34.61761, 49.41941))
plot(res$geometry, add = TRUE, pch = 19, cex = 0.05)
}
}
