% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{lagos_ingest}
\alias{lagos_ingest}
\title{Ingest LAGOSUS flat files}
\usage{
lagos_ingest(
  locus_version = NA,
  locus_folder = NA,
  limno_version = NA,
  limno_folder = NA,
  geo_version = NA,
  geo_folder = NA,
  depth_version = NA,
  depth_folder = NA
)
}
\arguments{
\item{locus_version}{character LAGOSUS database version string}

\item{locus_folder}{file.path to locus export folder}

\item{limno_version}{character LAGOSUS database version string}

\item{limno_folder}{file.path to limno export folder}

\item{geo_version}{character LAGOSUS database version string}

\item{geo_folder}{file.path to geo export folder}

\item{depth_version}{character LAGOSUS database version string}

\item{depth_folder}{file.path to depth export folder}
}
\description{
Ingest LAGOSUS data from component flat files
}
\examples{
\dontrun{
download_path <- "~/"
# download_path <- "../../../"

lg <- lagos_ingest(
 locus_version = "1.0",
 locus_folder = paste0(download_path, "Downloads/LOCUS_v1.0"))

lg <- lagos_ingest(
 limno_version = "2.1",
 limno_folder = paste0(download_path, "Downloads/LIMNO_v2.1/Final exports"))

lg <- lagos_ingest(
 depth_version = "0.1",
 depth_folder = paste0(download_path, "Downloads/DEPTH_v0.1"))
}
}
