% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{package}
\name{LAGOSUS-package}
\alias{LAGOSUS-package}
\alias{LAGOSUS}
\title{R interface to the LAGOSUS lakes database}
\description{
Interface to the LAGOSUS lakes database
}
\author{
\email{stachel2@msu.edu}
}
