% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{lagosus_compile}
\alias{lagosus_compile}
\title{Compile LAGOSUS data}
\usage{
lagosus_compile(
  locus_version = NA,
  locus_folder = NA,
  locus_overwrite = FALSE,
  limno_version = NA,
  limno_folder = NA,
  limno_overwrite = FALSE,
  geo_version = NA,
  geo_folder = NA,
  geo_overwrite = FALSE,
  depth_version = NA,
  depth_folder = NA,
  depth_overwrite = FALSE,
  dest_folder = NA
)
}
\arguments{
\item{locus_version}{character LAGOSUS database version string}

\item{locus_folder}{file.path to locus export folder}

\item{locus_overwrite}{logical overwrite existing data?}

\item{limno_version}{character LAGOSUS database version string}

\item{limno_folder}{file.path to limno export folder}

\item{limno_overwrite}{logical overwrite existing data?}

\item{geo_version}{character LAGOSUS database version string}

\item{geo_folder}{file.path to geo export folder}

\item{geo_overwrite}{logical overwrite existing data?}

\item{depth_version}{character LAGOSUS database version string}

\item{depth_folder}{file.path to depth export folder
#'@param depth_overwrite logical overwrite existing data?}

\item{dest_folder}{file.path optional will default to the location returned
by \code{\link[rappdirs]{user_data_dir}}}
}
\description{
Compile LAGOSUS data from component flat files. This function should rarely be called directly outside of manual updating of source data.
}
\examples{
\dontrun{
lagosus_compile(
 locus_version = "0",
 locus_folder = "~/Downloads/LAGOS-US-LOCUS-EXPORT",
 depth_version = "0",
 depth_folder = "~/Downloads/LAGOS-US-DEPTH-EXPORT",
 dest_folder = lagos_path())
}

}
