% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagosus_get}
\alias{lagosus_get}
\title{Retrieve LAGOSNE flat files}
\usage{
lagosus_get(
  version = lagosne_version(),
  overwrite = FALSE,
  dest_folder = tempdir()
)
}
\arguments{
\item{version}{character LAGOSNE database version string}

\item{overwrite}{logical overwrite existing data for the specified version}

\item{dest_folder}{file.path to save data. Default to a temporary folder.
Recommended to set to LAGOSNE:::lagosus_path() so that data persists between
R sessions.}
}
\description{
Retrieve LAGOSNE flat files from EDI.
}
\examples{
\dontrun{
# default to latest version
lagosus_get(dest_folder = LAGOSUS:::lagosus_path())

# get previous version(s)
# - recommended to install corresponding package version
# - See 'Legacy Versions' section of the README for instructions
}
}
