% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{lagosus_load}
\alias{lagosus_load}
\title{Load LAGOSUS data}
\usage{
lagosus_load(modules = NULL, versions = NULL, geo_tables = NA, fpath = NA)
}
\arguments{
\item{modules}{character vector of module names. One or more of "locus", "limno", "geo", or "depth.}

\item{versions}{character LAGOSUS database version strings. Defaults to \code{\link{lagosus_version}}}

\item{fpath}{file.path to LAGOSUS data store. Defaults to lagosus_path()}
}
\description{
Load LAGOSUS data from local system files
}
\examples{
\dontrun{
# works
lg <- lagosus_load("locus")
lg <- lagosus_load(modules = c("locus", "depth"), versions = c(0, 1))
lg <- lagosus_load(modules = c("locus", "depth"))
lg <- lagosus_load(modules = c("locus", "depth"), versions = c(0, NA))

# errors
# lg <- lagosus_load()
# lg <- lagosus_load(modules = c("locus", "depth"), versions = c(0))
}
}
