% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\name{lagosus_version}
\alias{lagosus_version}
\title{Latest LAGOSUS module versions}
\usage{
lagosus_version()
}
\description{
Latest LAGOSUS module versions
}
\details{
Defaults to the latest release version of each data module. Can be
overridden using the \code{LAGOSUS_LOCUS_VER}, \code{LAGOSUS_LIMNO_VER},
\code{LAGOSUS_GEO_VER}, or \code{LAGOSUS_DEPTH_VER} environment variables; see example
below. Using different environment variables for the same module in a single
interactive session will not invalidate the \code{lagosus_load} cache. Manually
specify module version numbers if needing to load multiple different module
versions.
}
\examples{
\dontrun{
# use defaults
lagosus_version()

# custom setting
Sys.setenv(LAGOSUS_DEPTH_VER = 0)
lagosus_version()

# revert to defaults
Sys.setenv(LAGOSUS_DEPTH_VER = "")
lagosus_version()
}
}
