% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{lake_info}
\alias{lake_info}
\title{Get a summary of lake attribute information}
\usage{
lake_info(
  lagoslakeid = NA,
  name = NA,
  state = NA,
  lg = lagosus_load("locus"),
  max_distance = 0
)
}
\arguments{
\item{lagoslakeid}{numeric lake id}

\item{name}{character lake name not caps sensitive}

\item{state}{character state name not caps sensitive}

\item{lg}{list of data.frames. output of \code{\link[LAGOSUS]{lagosus_load}}.}

\item{max_distance}{passed to \code{agrepl} "all" parameter
to control fuzzy matching lake name. Defaults to exact matching.}
}
\description{
Return lake attribute information from its name and state or its lagoslakeid.
}
\examples{
\dontrun{
lg <- lagosus_load("locus")

lake_info(lagoslakeid = 4314, lg = lg)
lake_info(lagoslakeid = c(21864, 2317))
lake_info(lagoslakeid = c(1441))
lake_info(lagoslakeid = c(125428, 1441), lg = lg)
lake_info(lagoslakeid = c(4686, 8016), lg = lg)

# fuzzy matching to lake name
lake_info(name = "Duck Lake", state = "Michigan", lg = lg)
# exact matching to lake name
lake_info(name = "Duck Lake", state = "Michigan",
          max.distance = list(all = 0), lg = lg)

lake_info(name = "Sunapee Lake", state = "New Hampshire", lg = lg)
lake_info(name = c("Sunapee Lake", "Oneida Lake"),
              state = c("New Hampshire", "New York"),
              max.distance = list(all = 0), lg = lg)
}
}
