% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{lake_link}
\alias{lake_link}
\title{LAGOSUS Identifier links}
\description{
Includes single or multiple identifiers per lake from other commonly used
national-scale data products including: Water Quality Portal (WQP),
Geographic Names Information System (GNIS), NHD MR Plusv2, LAGOS-NE, and
the EPA National Lake Assessment surveys from 2007 and 2012.
}
\details{
\tabular{lllll}{
\bold{Variable name} \tab \bold{Variable name group} \tab \bold{Lagosne variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                      \tab lagoslakeid                      \tab lagoslakeid \tab unique lake identifier developed by LAGOS-US                                                                                                                \tab                 \cr
lake_nhdid                       \tab nhdid                            \tab nhdid       \tab the unique Permanent_Identifier from the NHD for each LAGOS lake                                                                                            \tab                 \cr
lake_reachcode                   \tab reachcode                        \tab             \tab reach code assigned to the water feature by the NHD                                                                                                         \tab                 \cr
lake_namegnis                    \tab namegnis                         \tab gnis_name   \tab lake name from the GNIS database                                                                                                                            \tab                 \cr
lake_namelagos                   \tab namelagos                        \tab lagosname1  \tab lake name from a combination of data sources; examples are GNIS, WQP, etc.                                                                                  \tab                 \cr
lake_county                      \tab county                           \tab             \tab name of county containing the focal lake polygon centroid                                                                                                   \tab                 \cr
lake_countyfips                  \tab countyfips                       \tab             \tab Federal Information Processing Standard, FIPS, code for the county containing the focal lake polygon centroid                                               \tab                 \cr
lake_lat_decdeg                  \tab lat_decdeg                       \tab nhd_lat     \tab latitude of central point inside the NHD lake polygon in decimal degrees; NAD83 projection                                                                  \tab decimal degree  \cr
lake_lon_decdeg                  \tab lon_decdeg                       \tab nhd_long    \tab longitude of  central point inside the NHD lake polygon in decimal degrees; NAD83 projection                                                                \tab decimal degree  \cr
lake_centroidstate               \tab centroidstate                    \tab             \tab two-letter postal abbreviation of the state containing the lake central point                                                                               \tab                 \cr
nhdhr_area_sqkm                  \tab nhdhr_area_sqkm                  \tab             \tab area of areal feature based on Albers Equal Area, as assigned in NHD                                                                                        \tab square kilometer\cr
nhdhr_fdate                      \tab nhdhr_fdate                      \tab             \tab date of last feature modification in National Hydrography Dataset                                                                                           \tab                 \cr
nhdhr_gnisid                     \tab nhdhr_gnisid                     \tab             \tab unique identifier assigned by GNIS                                                                                                                          \tab                 \cr
lagosus_legacysiteid             \tab lagosus_legacysiteid             \tab             \tab unique site or lake identifier from the original LAGOS-US water quality source dataset--this is not standardized and each source dataset has its own system \tab                 \cr
lagosus_legacysitelabel          \tab lagosus_legacysitelabel          \tab             \tab column name or field name used by the source dataset to label the legacy lake identifier                                                                    \tab                 \cr
lagosus_legacyprogram            \tab lagosus_legacyprogram            \tab             \tab program name assigned in LAGOS-US to identify the source dataset provider                                                                                   \tab                 \cr
wqp_monitoringlocationidentifier \tab wqp_monitoringlocationidentifier \tab             \tab a designator used to describe the unique name, number, or code assigned to identify the monitoring location                                                 \tab                 \cr
wqp_monitoringlocationname       \tab wqp_monitoringlocationname       \tab             \tab the designator specified by the sampling organization for the site at which sampling or other activities are conducted                                      \tab                 \cr
wqp_providername                 \tab wqp_providername                 \tab             \tab source system that provided data to the Water Quality Portal; examples are NWIS, STORET, STEWARDS, etc.                                                     \tab                 \cr
nhdplusv2_comid                  \tab nhdplusv2_comid                  \tab             \tab common identifier of the NHD Waterbody feature                                                                                                              \tab                 \cr
nhdplusv2_reachcode              \tab nhdplusv2_reachcode              \tab             \tab NHDplus v2 reach code assigned to the water feature.                                                                                                        \tab                 \cr
nhdplusv2_area_sqkm              \tab nhdplusv2_area_sqkm              \tab             \tab feature area in square kilometers, as assigned by NHD                                                                                                       \tab square kilometer\cr
lagosne_lagoslakeid              \tab lagosne_lagoslakeid              \tab             \tab unique lake identifier developed for LAGOS-NE                                                                                                               \tab                 \cr
lagosne_legacysiteid             \tab lagosne_legacysiteid             \tab             \tab unique site/lake identifier from the original LAGOS-NE water quality source dataset--this is not standardized and each source dataset has its own system    \tab                 \cr
nla2007_siteid                   \tab nla2007_siteid                   \tab             \tab lake site identifier from National Lakes Assessment 2007 survey                                                                                             \tab                 \cr
nla2012_siteid                   \tab nla2012_siteid                   \tab             \tab lake site identifier from National Lakes Assessment 2012 survey                                                                                             \tab                 \cr
nhdplusv2_lakes_n                \tab nhdplusv2_lakes_n                \tab             \tab count of NHDPlusV2 polygons matching the LAGOS-US lagoslakeid                                                                                               \tab count           \cr
lagosne_lakes_n                  \tab lagosne_lakes_n                  \tab             \tab count of LAGOS-NE lagoslakeids matching the LAGOS-US lagoslakeid                                                                                            \tab count           \cr
wqp_sites_n                      \tab wqp_sites_n                      \tab             \tab count of WQP sites within the LAGOS-US lake                                                                                                                 \tab count           \cr
lagosus_legacyids_n              \tab lagosus_legacyids_n              \tab             \tab count of all LAGOS-US legacy ids linked to the lagoslakeid in this table; these ids include, for example, original sampling site identifiers                \tab count
}
}
\keyword{datasets}
