% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{lake_watersheds}
\alias{lake_watersheds}
\alias{watersheds}
\title{LAGOSUS Lake watersheds}
\description{
Includes identifiers, location, and geometry for the calculated watersheds (WS) and network
watersheds (NWS).
}
\details{
\tabular{lllll}{
\bold{Variable name} \tab \bold{Variable name group} \tab \bold{Lagosne variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                    \tab lagoslakeid              \tab lagoslakeid    \tab unique lake identifier developed by LAGOS-US                                                                                                                  \tab               \cr
ws_zoneid                      \tab zoneid                   \tab iws_zoneid     \tab unique WS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                   \tab               \cr
nws_zoneid                     \tab zoneid                   \tab                \tab unique NWS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                  \tab               \cr
ws_subtype                     \tab subtype                  \tab                \tab watershed subtype class for the watershed; LC equals local catchment, DWS equals drainage-watershed, IDWS equals inter-drainage-lake watershed                \tab               \cr
ws_equalsnws                   \tab equalsnws                \tab                \tab flag indicating that the ws is equivalent to the nws                                                                                                          \tab               \cr
ws_onlandborder                \tab onlandborder             \tab                \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                               \tab               \cr
ws_oncoast                     \tab oncoast                  \tab                \tab flag indicating the zone is adjacent to an ocean or Great Lakes coast                                                                                         \tab               \cr
ws_inusa_pct                   \tab inusa_pct                \tab iws_pct_in_usa \tab percent of the zone polygon that is within the US border                                                                                                      \tab percent       \cr
ws_includeshu4inlet            \tab includeshu4inlet         \tab                \tab flag indicating the WS extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                         \tab               \cr
ws_ismultipart                 \tab ismultipart              \tab                \tab flag indicating that the focal zone polygon is multipart                                                                                                      \tab               \cr
ws_sliverflag                  \tab sliverflag               \tab                \tab flag indicating extreme cases where the watershed is comprised of slivers due to flat or constrained terrain that may not accurately reflect surface drainage \tab               \cr
nws_onlandborder               \tab onlandborder             \tab                \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                               \tab               \cr
nws_oncoast                    \tab oncoast                  \tab                \tab flag indicating the zone is adjacent to an ocean or Great Lakes coast                                                                                         \tab               \cr
nws_inusa_pct                  \tab inusa_pct                \tab                \tab percent of the zone polygon that is within the US border                                                                                                      \tab percent       \cr
nws_includeshu4inlet           \tab includeshu4inlet         \tab                \tab flag indicating the NWS extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                        \tab               \cr
nws_ismultipart                \tab ismultipart              \tab                \tab flag indicating that the focal zone polygon is multipart                                                                                                      \tab               \cr
ws_states                      \tab states                   \tab                \tab two-letter postal abbreviations of all states intersecting the zone polygon                                                                                   \tab               \cr
ws_focallakewaterarea_ha       \tab focallakewaterarea_ha    \tab iws_lakeareaha \tab surface area of focal lake for which the WS is calculated                                                                                                     \tab hectare       \cr
ws_area_ha                     \tab area_ha                  \tab iws_ha         \tab area of zone polygon                                                                                                                                          \tab hectare       \cr
ws_perimeter_m                 \tab perimeter_m              \tab iws_perimkm    \tab perimeter of zone polygon including perimeter of any internal holes                                                                                           \tab meter         \cr
ws_lake_arearatio              \tab lake_arearatio           \tab                \tab ratio between WS area and lake water area                                                                                                                     \tab               \cr
ws_mbgconhull_length_m         \tab mbgconhull_length_m      \tab                \tab the maximum possible distance between any two vertices of the minimum bounding geometry convex hull that encloses the WS polygon                              \tab meter         \cr
ws_mbgconhull_width_m          \tab mbgconhull_width_m       \tab                \tab the maximum width of the WS convex hull polygon along an axis perpendicular to the axis defined by the convex hull length                                     \tab meter         \cr
ws_mbgconhull_orientation_deg  \tab mbgconhull_orientation_m \tab                \tab orientation of the line defining the convex hull length; calculated as decimal degrees clockwise from north                                                   \tab degree        \cr
ws_meanwidth_m                 \tab meanwidth_m              \tab                \tab mean width of the WS polygon; calculated as ws_area_ha converted to hectares and divided by ws_mbgconhull_length_m                                            \tab meter         \cr
ws_lat_decdeg                  \tab lat_decdeg               \tab iws_lat        \tab latitude of the focal lake central point in decimal degrees; NAD83 projection                                                                                 \tab decimaldegree \cr
ws_lon_decdeg                  \tab lon_decdeg               \tab iws_lon        \tab longitude of the focal lake central point in decimal degrees; NAD83 projection                                                                                \tab decimaldegree \cr
nws_states                     \tab states                   \tab                \tab two-letter postal abbreviations of all states intersecting the zone polygon                                                                                   \tab               \cr
nws_focallakewaterarea_ha      \tab focallakewaterarea_ha    \tab                \tab surface area of focal lake for which the NWS is calculated                                                                                                    \tab hectare       \cr
nws_area_ha                    \tab area_ha                  \tab                \tab area of zone polygon                                                                                                                                          \tab hectare       \cr
nws_perimeter_m                \tab perimeter_m              \tab                \tab perimeter of zone polygon including perimeters of any internal holes                                                                                          \tab meter         \cr
nws_lake_arearatio             \tab lake_arearatio           \tab                \tab ratio between NWS area and lake water area                                                                                                                    \tab               \cr
nws_mbgconhull_length_m        \tab mbgconhull_length_m      \tab                \tab the maximum possible distance between any two vertices of the minimum bounding geometry convex hull that encloses the NWS polygon                             \tab meter         \cr
nws_mbgconhull_width_m         \tab mbgconhull_width_m       \tab                \tab the maximum width of the NWS convex hull polygon along an axis perpendicular to the axis defined by the convex hull length                                    \tab meter         \cr
nws_mbgconhull_orientation_deg \tab mbgconhull_orientation_m \tab                \tab orientation of the line defining the convex hull length; calculated as decimal degrees clockwise from north                                                   \tab degree        \cr
nws_meanwidth_m                \tab meanwidth_m              \tab                \tab mean width of the NWS polygon; calculated as nws_area_ha converted to hectares and divided by nws_mbgconhull_length_m                                         \tab meter         \cr
nws_lat_decdeg                 \tab lat_decdeg               \tab                \tab latitude of the focal lake central point in decimal degrees; NAD83 projection                                                                                 \tab decimal degree\cr
nws_lon_decdeg                 \tab lon_decdeg               \tab                \tab longitude of the focal lake central point in decimal degrees; NAD83 projection                                                                                \tab decimal degree
}
}
\keyword{datasets}
