% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{locus_characteristics}
\alias{locus_characteristics}
\title{LAGOSUS Lake characteristics}
\description{
Includes derived metrics for each lake such as lake geometry (e.g., area,
 perimeter, shape metrics), lake connectivity (e.g., lake connectivity
 class, number of upstream lakes), and whether the lake is located within an
 area that was glaciated during the Late Wisconsin glaciation.
}
\details{
\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                     \tab unique lake identifier developed by LAGOS-US                                                                                                                                    \tab        \cr
  lake_waterarea_ha               \tab surface area of lake waterbody polygon from NHD (excludes islands)                                                                                                              \tab hectare\cr
  lake_totalarea_ha               \tab surface area within outer boundary of lake waterbody polygon from NHD (includes islands)                                                                                        \tab hectare\cr
  lake_islandarea_ha              \tab surface area of islands within outer boundary of lake waterbody polygon from NHD                                                                                                \tab hectare\cr
  lake_perimeter_m                \tab perimeter of outer boundary of lake waterbody polygon from the NHD (excludes islands)                                                                                           \tab meter  \cr
  lake_islandperimeter_m          \tab perimeter of islands within the lake waterbody polygon from the NHD                                                                                                             \tab meter  \cr
  lake_shorelinedevfactor         \tab shoreline development factor = lake_perimeter / (2*?(?*waterarea))                                                                                                              \tab        \cr
  lake_mbgconhull_length_m        \tab the longest distance entirely within the lake polygon between any two vertices of the minimum bounding geometry convex hull that encloses the lake polygon                      \tab meter  \cr
  lake_mbgconhull_width_m         \tab the shortest distance entirely within the lake polygon between any two vertices of the minimum bounding geometry convex hull that encloses the lake polygon                     \tab meter  \cr
  lake_mbgconhull_orientation_deg \tab orientation (from 0 to 180 decimal degrees clockwise from north) of the longest line connecting two vertices entirely within both the convex hull and the lake polygon          \tab degree \cr
  lake_mbgrect_length_m           \tab length of the minimum bounding geometry rectangle by area enclosing the lake polygon                                                                                            \tab meter  \cr
  lake_mbgrect_width_m            \tab width of the minimum bounding geometry rectangle by area enclosing the lake polygon                                                                                             \tab meter  \cr
  lake_mbgrect_arearatio          \tab ratio between lake_waterarea_ha (* 10000 m^2/ha) and the area of the minimum bounding rectangle by area of the lake polygon; value near 1 indicates lake shape is rectangular
 \tab        \cr
  lake_meanwidth_m                \tab mean width of the lake calculated as lake water area (lake_area_ha* 10000 m^2/ha) divided by the convex hull length of the focal lake polygon (lake_mbgconhull_length_m)        \tab meter  \cr
  lake_connectivity_class         \tab hydrologic connectivity class of the focal lake determined from the NHD network considering both permanent and intermittent/ephemeral flow                                      \tab        \cr
  lake_connectivity_fluctuates    \tab indicates whether the lake's connectivity classification depends on non-permanent flow                                                                                          \tab        \cr
  lake_connectivity_permanent     \tab hydrologic connectivity class of the focal lake determined from the NHD network considering only permanent flow                                                                 \tab        \cr
  lake_lakes1ha_upstream_ha       \tab total area of lakes >= 1 ha upstream of the focal lake, connected via surface streams                                                                                           \tab hectare\cr
  lake_lakes4ha_upstream_ha       \tab total area of lakes >= 4 ha upstream of the focal lake, connected via surface streams                                                                                           \tab hectare\cr
  lake_lakes10ha_upstream_ha      \tab total area of lakes >= 10 ha upstream of the focal lake, connected via surface streams                                                                                          \tab hectare\cr
  lake_lakes1ha_upstream_n        \tab count of lakes >= 1 ha upstream of the focal lake, connected via surface streams                                                                                                \tab count  \cr
  lake_lakes4ha_upstream_n        \tab count of lakes >= 4 ha upstream of the focal lake, connected via surface streams                                                                                                \tab count  \cr
  lake_lakes10ha_upstream_n       \tab count of lakes >= 10 ha upstream of the focal lake, connected via surface streams                                                                                               \tab count  \cr
  lake_glaciatedlatewisc          \tab glaciation status during the Late Wisconsin glaciation (status = Glaciated if any part of the lake polygon is inside the glacial extent)                                        \tab        
}
}
\keyword{datasets}
