% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{locus_information}
\alias{locus_information}
\title{LAGOSUS Lake information}
\description{
LAGOSUS Lake information
}
\details{
\tabular{ll}{
lagoslakeid        \tab LAGOS sourced unique identifier for the focal lake \cr
lake_nhdid         \tab the unique 'Permanent_identifier' from the NHD for each LAGOS lake \cr
 lake_nhdfcode      \tab NHD five-digit integer code with feature code plus additional characteristics/values                                            \cr
 lake_nhdftype      \tab NHD three-digit integer code providing a unique identifier of feature type for the waterbody                                    \cr
 lake_reachcode     \tab NHD reachcode for stream reach intersecting the lake from NHD stream network                                                    \cr
 lake_namegnis      \tab lake name from the gnis database                                                                                                \cr
 lake_namelagos     \tab lake name from a combination of data sources (gnis, wqp, etc.)                                                                  \cr
 ♦                  \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                 \cr
 lake_ismultipart   \tab flag indicating that the focal lake polygon is multipart                                                                        \cr
 lake_missingws     \tab flag indicating that the lake's watershed (ws or nws) was not delineated                                                        \cr
 lake_shapeflag     \tab flag indicating lake polygon shape is excessively angular (e.g., triangle, rectangle) or elongate (very thin relative to length)\cr
 lake_lat_decdeg    \tab latitude of centroid of the NHD lake polygon in decimal degrees (NAD83)                                                         \cr
 lake_lon_decdeg    \tab longitude of  centroid of the NHD lake polygon in decimal degrees (NAD83)                                                       \cr
 lake_centroidstate \tab abbreviation of state containing the lake centroid                                                                              \cr
 lake_states        \tab abbreviation(s) of state(s) intersecting the zone polygon                                                                       \cr
 lake_huc12         \tab HUC12 of the WBD HU12 watershed containing the lake                                                                             \cr
 buff100_zoneid     \tab unique 100m buffer zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS                                   \cr
 buff500_zoneid     \tab unique 500m buffer zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS                                   \cr
 ws_zoneid          \tab unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS                                               \cr
 nws_zoneid         \tab unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS                                               \cr
 hu12_zoneid        \tab unique HU12 zone identifier assigned by LAGOS for zone containing the focal lake                                                \cr
 hu8_zoneid         \tab unique HU8 zone identifier assigned by LAGOS for zone containing the focal lake                                                 \cr
 hu4_zoneid         \tab unique HU4 zone identifier assigned by LAGOS for zone containing the focal lake                                                 \cr
 county_zoneid      \tab unique county zone identifier assigned by LAGOS for zone containing the focal lake                                              \cr
 state_zoneid       \tab unique state zone identifier assigned by LAGOS for zone containing the focal lake                                               \cr
 epanutr_zoneid     \tab unique EPA Nutrient Ecoregion zone identifier assigned by LAGOS for zone containing the focal lake                              \cr
 omernik3_zoneid    \tab unique Omernik Level III Ecoregion zone identifier assigned by LAGOS for zone containing the focal lake                         \cr
 wwf_zoneid         \tab unique WWF zone identifier assigned by LAGOS for zone containing the focal lake                                                 \cr
 mlra_zoneid        \tab unique Major Land Resource Area zone identifier assigned by LAGOS for zone containing the focal lake                            \cr
 bailey_zoneid      \tab unique Bailey Ecoregion zone identifier assigned by LAGOS for zone containing the focal lake                                    \cr
 neon_zoneid        \tab unique NEON zone identifier assigned by LAGOS for zone containing the focal lake
}
}
\keyword{datasets}
