% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{locus_link}
\alias{locus_link}
\title{LAGOSUS Identifier links}
\description{
LAGOSUS Identifier links
}
\details{
\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
nhdhr_gnisid                     \tab unique identifier assigned by GNIS                                                                                                                        \tab                 \cr
  lagoslakeid                      \tab unique lake identifier developed by LAGOS-US                                                                                                              \tab                 \cr
  lake_nhdid                       \tab the unique 'Permanent_identifier' from the NHD for each LAGOS lake                                                                                        \tab                 \cr
  lake_reachcode                   \tab reach code assigned to the water feature by the NHD                                                                                                       \tab                 \cr
  lake_namegnis                    \tab lake name from the gnis database                                                                                                                          \tab                 \cr
  lake_namelagos                   \tab lake name from a combination of data sources (gnis, wqp, etc.)                                                                                            \tab                 \cr
  lake_county                      \tab NULLme of county containing the focal lake polygon centroid                                                                                               \tab                 \cr
  lake_countyfips                  \tab Federal Information Processing Standard (FIPS) code for the county containing the focal lake polygon centroid                                             \tab                 \cr
  lake_lat_decdeg                  \tab latitude of centroid of the NHD lake polygon in decimal degrees (NAD83)                                                                                   \tab decimal degree  \cr
  lake_lon_decdeg                  \tab longitude of  centroid of the NHD lake polygon in decimal degrees (NAD83)                                                                                 \tab decimal degree  \cr
  lake_centroidstate               \tab abbreviation of state containing the lake centroid                                                                                                        \tab                 \cr
  nhdhr_area_sqkm                  \tab NHD feature area in square kilometers                                                                                                                     \tab square kilometer\cr
  nhdhr_fdate                      \tab date of last feature modification in National Hydrography Dataset                                                                                         \tab                 \cr
  lagosus_legacysiteid             \tab Unique site/lake identifier from the original LAGOS-US water quality source dataset. This is not standardized and each source dataset has its own system. \tab                 \cr
  lagosus_legacysitelabel          \tab the column/field name used by the source dataset to label the legacy lake identifier                                                                      \tab                 \cr
  lagosus_legacyprogram            \tab the program name assigned in LAGOS_US to identify the source dataset provider                                                                             \tab                 \cr
  wqp_monitoringlocationidentifier \tab a designator used to describe the unique name, number, or code assigned to identify the monitoring location                                               \tab                 \cr
  wqp_monitoringlocationname       \tab the designator specified by the sampling organization for the site at which sampling or other activities are conducted                                    \tab                 \cr
  wqp_providername                 \tab the source system that provided data to the Water Quality Portal (NWIS, STORET, STEWARDS, etc)                                                            \tab                 \cr
  nhdplusv2_comid                  \tab common identifier of the NHD Waterbody feature                                                                                                            \tab                 \cr
  nhdplusv2_reachcode              \tab NHDplus v2 reach code assigned to the water feature.                                                                                                      \tab                 \cr
  nhdplusv2_area_sqkm              \tab NHD feature area in square kilometers                                                                                                                     \tab square kilometer\cr
  lagosne_lagoslakeid              \tab unique lake identifier developed for LAGOS-NE                                                                                                             \tab                 \cr
  lagosne_legacysiteid             \tab Unique site/lake identifier from the original LAGOS-NE water quality source dataset. This is not standardized and each source dataset has its own system. \tab                 \cr
  nla2007_siteid                   \tab lake site identifier from National Lakes Assessment 2007 survey                                                                                           \tab                 \cr
  nla2012_siteid                   \tab lake site identifier from National Lakes Assessment 2012 survey                                                                                           \tab                 \cr
  nhdplusv2_lakes_n                \tab count of WQP sites within the LAGOS-US lake                                                                                                               \tab count           \cr
  lagosne_lakes_n                  \tab count of NHDPlusV2 polygons matching the LAGOS-US lagoslakeid                                                                                             \tab count           \cr
  wqp_sites_n                      \tab count of LAGOS-NE lagoslakeids matching the LAGOS-US lagoslakeid                                                                                          \tab count           \cr
  lagosus_legacyids_n              \tab count of all LAGOS-US legacy (original) sampling site identifiers linked to the lagoslakeid in this table                                                 \tab count
}
}
\keyword{datasets}
