% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{locus_watersheds}
\alias{locus_watersheds}
\alias{watersheds}
\title{LAGOSUS Lake watersheds}
\description{
LAGOSUS Lake watersheds

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                    \tab unique lake identifier developed by LAGOS-US                                                                                                                          \tab               \cr
  ws_zoneid                      \tab unique watershed identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid                                                                            \tab               \cr
  nws_zoneid                     \tab unique network watershed identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid                                                                    \tab               \cr
  ws_subtype                     \tab LAGOS watershed subtype class for the watershed (LC=local catchment, DWS=drainage-watershed, IDWS=inter-drainage-lake watershed)                                      \tab               \cr
  ws_equalsnws                   \tab flag indicating that the ws is equivalent to the nws (=Y where there are no upstream lakes > 10 ha and where lake connectivity class does not equal drainagelk)       \tab               \cr
  ws_onlandborder                \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                                       \tab               \cr
  ws_oncoast                     \tab flag indicating the zone is adjacent to the coastline                                                                                                                 \tab               \cr
  ws_inusa_pct                   \tab percent of the zone (i.e., polygon) that is within the US border                                                                                                      \tab percent       \cr
  ws_includeshu4inlet            \tab flag indicating the ws extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                                 \tab               \cr
  ws_ismultipart                 \tab flag indicating that the focal zone polygon is multipart                                                                                                              \tab               \cr
  ws_sliverflag                  \tab flag indicating extreme cases where the watershed is comprised of slivers due to flat or constrained terrain that may not accurately reflect surface drainage         \tab               \cr
  nws_onlandborder               \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                                       \tab               \cr
  nws_oncoast                    \tab flag indicating the zone is adjacent to the coastline                                                                                                                 \tab               \cr
  nws_inusa_pct                  \tab percent of the zone (i.e., polygon) that is within the US border                                                                                                      \tab percent       \cr
  nws_includeshu4inlet           \tab flag indicating the nws extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                                \tab               \cr
  nws_ismultipart                \tab flag indicating that the focal zone polygon is multipart                                                                                                              \tab               \cr
  ws_states                      \tab abbreviation(s) of state(s) intersecting the zone polygon                                                                                                             \tab               \cr
  ws_focallakewaterarea_ha       \tab surface area of focal lake for which the watershed (WS) is calculated                                                                                                 \tab hectare       \cr
  ws_area_ha                     \tab area of zone polygon                                                                                                                                                  \tab hectare       \cr
  ws_perimeter_m                 \tab perimeter of zone polygon including perimeters of any internal 'holes'                                                                                                \tab meter         \cr
  ws_lake_arearatio              \tab ratio between watershed area and lake water area                                                                                                                      \tab               \cr
  ws_mbgconhull_length_m         \tab the longest distance entirely within the ws polygon between any two vertices of the minimum bounding geometry convex hull that encloses the ws polygon                \tab meter         \cr
  ws_mbgconhull_width_m          \tab the shortest distance entirely within the ws polygon between any two vertices of the minimum bounding geometry convex hull that encloses the ws polygon               \tab meter         \cr
  ws_mbgconhull_orientation_deg  \tab orientation (from 0 to 180 decimal degrees clockwise from north) of the longest line connecting two vertices entirely within both the convex hull and the ws polygon  \tab degree        \cr
  ws_meanwidth_m                 \tab mean width of the ws calculated as ws area (ws_area_ha* 10000 m^2/ha) divided by the convex hull length of the focal ws polygon (ws_mbgconhull_length_m)              \tab meter         \cr
  ws_lat_decdeg                  \tab latitude of the centroid of the focal lake polygon in decimal degrees (NAD83)                                                                                         \tab decimaldegree \cr
  ws_lon_decdeg                  \tab longitude of the centroid of the focal lake polygon in decimal degrees (NAD83)                                                                                        \tab decimaldegree \cr
  nws_states                     \tab abbreviation(s) of state(s) intersecting the zone polygon                                                                                                             \tab               \cr
  nws_focallakewaterarea_ha      \tab surface area of focal lake for which the network watershed (NWS) is calculated                                                                                        \tab hectare       \cr
  nws_area_ha                    \tab area of zone polygon                                                                                                                                                  \tab hectare       \cr
  nws_perimeter_m                \tab perimeter of zone polygon including perimeters of any internal 'holes'                                                                                                \tab meter         \cr
  nws_lake_arearatio             \tab ratio between network watershed area and lake water area                                                                                                              \tab               \cr
  nws_mbgconhull_length_m        \tab the longest distance entirely within the nws polygon between any two vertices of the minimum bounding geometry convex hull that encloses the nws polygon              \tab meter         \cr
  nws_mbgconhull_width_m         \tab the shortest distance entirely within the nws polygon between any two vertices of the minimum bounding geometry convex hull that encloses the nws polygon             \tab meter         \cr
  nws_mbgconhull_orientation_deg \tab orientation (from 0 to 180 decimal degrees clockwise from north) of the longest line connecting two vertices entirely within both the convex hull and the nws polygon \tab degree        \cr
  nws_meanwidth_m                \tab mean width of the nws calculated as nws area (nws_area_ha* 10000 m^2/ha) divided by the convex hull length of the focal nws polygon (nws_mbgconhull_length_m)         \tab meter         \cr
  nws_lat_decdeg                 \tab latitude of the centroid of the focal lake polygon in decimal degrees (NAD83)                                                                                         \tab decimal degree\cr
  nws_lon_decdeg                 \tab longitude of the centroid of the focal lake polygon in decimal degrees (NAD83)                                                                                        \tab decimal degree
}
}
\keyword{datasets}
