% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{locus_ws}
\alias{locus_ws}
\alias{watersheds}
\title{LAGOSUS Lake watersheds}
\description{
\tabular{ll}{
  lagoslakeid                    \tab LAGOS sourced unique identifier for the focal lake of the zone                                                                                                                           \cr
  ws_zoneid                      \tab unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS                                                                                                        \cr
  nws_zoneid                     \tab unique zone identifier equivalent to the focal lake lagoslakeid assigned by LAGOS                                                                                                        \cr
  ws_subtype                     \tab LAGOS watershed sub-type classification for the watershed                                                                                                                                \cr
  ws_equalsnws                   \tab flag indicating that the focal zone polygon is multipart                                                                                                                                 \cr
  ws_onlandborder                \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                                                          \cr
  ws_oncoast                     \tab flag indicating the zone is adjacent to the coastline                                                                                                                                    \cr
  ws_inusa_pct                   \tab percent of the zone (i.e., polygon) that is within the US border                                                                                                                         \cr
  ws_includeshu4inlet            \tab flag indicating the ws extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                                                    \cr
  ws_ismultipart                 \tab flag indicating that the zonal polygon is multipart                                                                                                                                      \cr
  ws_sliver_flag                 \tab NA                                                                                                                                                                                       \cr
  nws_onlandborder               \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                                                          \cr
  nws_oncoast                    \tab flag indicating the zone is adjacent to the coastline                                                                                                                                    \cr
  nws_inusa_pct                  \tab percent of the zone (i.e., polygon) that is within the US border                                                                                                                         \cr
  nws_includeshu4inlet           \tab flag indicating the nws extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                                                   \cr
  nws_ismultipart                \tab flag indicating that the zonal polygon is multipart                                                                                                                                      \cr
  ws_states                      \tab abbreviation(s) of state(s) intersecting the zone polygon                                                                                                                                \cr
  ws_focallakewaterarea_ha       \tab surface area of focal lake for which the watershed (WS) is calculated                                                                                                                    \cr
  ws_area_ha                     \tab area of focal zone polygon                                                                                                                                                               \cr
  ws_perimeter_m                 \tab perimeter of the focal zone polygon including perimeters of any internal 'holes'                                                                                                         \cr
  ws_lake_arearatio              \tab NA                                                                                                                                                                                       \cr
  ws_mbgconhull_length_m         \tab the longest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon                                                                   \cr
  ws_mbgconhull_width_m          \tab the shortest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon                                                                  \cr
  ws_mbgconhull_orientation_deg  \tab orientation (from 0 to 180 degrees) of the line connecting the antipodal pairs defining the longest distance within the minimum bounding geometry convex hull enclosing the zonal polygon\cr
  ws_meanwidth_m                 \tab ws area (* 10000 m^2/ha) divided by the ws convex hull length of the focal zone polygon                                                                                                  \cr
  nws_states                     \tab abbreviation(s) of state(s) intersecting the zone polygon                                                                                                                                \cr
  nws_focallakewaterarea_ha      \tab surface area of focal lake for which the watershed (NWS) is calculated                                                                                                                   \cr
  nws_area_ha                    \tab area of zone polygon                                                                                                                                                                     \cr
  nws_perimeter_m                \tab perimeter of the focal zone polygon including perimeters of any internal 'holes'                                                                                                         \cr
  nws_lake_arearatio             \tab NA                                                                                                                                                                                       \cr
  nws_mbgconhull_length_m        \tab the longest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon                                                                   \cr
  nws_mbgconhull_width_m         \tab the shortest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon                                                                  \cr
  nws_mbgconhull_orientation_deg \tab orientation (from 0 to 180 degrees) of the line connecting the antipodal pairs defining the longest distance within the minimum bounding geometry convex hull enclosing the zonal polygon\cr
  nws_meanwidth_m                \tab nws area (* 10000 m^2/ha) divided by the nws convex hull length of the focal zone polygon
}
}
\keyword{datasets}
