% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{site_chemicalphysical}
\alias{site_chemicalphysical}
\title{LAGOSUS Chemical and physical data}
\description{
LAGOSUS Chemical and physical data

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                         \tab unique lake identifier developed by LAGOSUS                                                                                                                                                                  \tab NULL                       \cr
eventid_epi                         \tab unique event identifier that is a combination of programid, lakeid, and date for each sampling event in LAGOS                                                                                                \tab NULL                       \cr
site_legacyid                       \tab unique sample site identifier provided by the source of the data; this is not standardardized and each source dataset has its own system                                                                     \tab NULL                       \cr
program_id                          \tab unique integer program identifier assigned by LAGOSUS                                                                                                                                                        \tab NULL                       \cr
program_name                        \tab name of sampling program that collected the data                                                                                                                                                             \tab NULL                       \cr
event_date                          \tab date, as YYYY-MM-DD, when the lake was sampled by the source program                                                                                                                                         \tab YYYY-MM-DD                 \cr
site_id                             \tab unique sample site identifier created by the LAGOSUS                                                                                                                                                         \tab NULL                       \cr
site_lat_decdeg                     \tab latitude of the sample site                                                                                                                                                                                  \tab Decimal Degrees            \cr
site_lon_decdeg                     \tab longitude of the sample site                                                                                                                                                                                 \tab Decimal Degrees            \cr
program_type                        \tab type of program that provided the data                                                                                                                                                                       \tab NULL                       \cr
alk_ueql                            \tab alkalinity, microequivalents per liter                                                                                                                                                                       \tab Microequivalents per Liter \cr
alk_qualifier_legacyflag            \tab alkalinity data flags from the source program                                                                                                                                                                \tab NULL                       \cr
alk_qualifier_legacyflagdesc        \tab alkalinity data flag descriptions from the source program                                                                                                                                                    \tab NULL                       \cr
alk_censorcode                      \tab the LAGOSUS assigned censor code for alkalinity that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself           \tab NULL                       \cr
alk_detectionlimit_legacyvalue      \tab alkalinity detection limit used by the source program                                                                                                                                                        \tab NULL                       \cr
alk_detectionlimit_legacyunit       \tab alkalinity detection limit unit used by the source program                                                                                                                                                   \tab NULL                       \cr
alk_method                          \tab alkalinity analytical procedure, from a standards body if available                                                                                                                                          \tab NULL                       \cr
ca_mgl                              \tab calcium, dissolved, milligrams per liter                                                                                                                                                                     \tab Milligrams per Liter       \cr
ca_qualifier_legacyflag             \tab calcium, dissolved data flags from the source program                                                                                                                                                        \tab NULL                       \cr
ca_qualifier_legacyflagdesc         \tab calcium, dissolved data flag descriptions from the source program                                                                                                                                            \tab NULL                       \cr
ca_censorcode                       \tab the LAGOSUS assigned censor code for calcium, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself   \tab NULL                       \cr
ca_detectionlimit_legacyvalue       \tab calcium, dissolved detection limit used by the source program                                                                                                                                                \tab NULL                       \cr
ca_detectionlimit_legacyunit        \tab calcium, dissolved detection limit unit used by the source program                                                                                                                                           \tab NULL                       \cr
ca_method                           \tab calcium, dissolved analytical procedure, from a standards body if available                                                                                                                                  \tab NULL                       \cr
cl_mgl                              \tab chloride, dissolved, milligrams per liter                                                                                                                                                                    \tab Milligrams per Liter       \cr
cl_qualifier_legacyflag             \tab chloride, dissolved data flags from the source program                                                                                                                                                       \tab NULL                       \cr
cl_qualifier_legacyflagdesc         \tab chloride, dissolved data flag descriptions from the source program                                                                                                                                           \tab NULL                       \cr
cl_censorcode                       \tab the LAGOSUS assigned censor code for chloride, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself  \tab NULL                       \cr
cl_detectionlimit_legacyvalue       \tab chloride, dissolved detection limit used by the source program                                                                                                                                               \tab NULL                       \cr
cl_detectionlimit_legacyunit        \tab chloride, dissolved detection limit unit used by the source program                                                                                                                                          \tab NULL                       \cr
cl_method                           \tab chloride, dissolved analytical procedure, from a standards body if available                                                                                                                                 \tab NULL                       \cr
spcond_uscm                         \tab specific conductance, microsiemens per cm at 25C                                                                                                                                                             \tab Microsiemens per Centimeter\cr
spcond_qualifier_legacyflag         \tab specific conductance data flags from the source program                                                                                                                                                      \tab NULL                       \cr
spcond_qualifier_legacyflagdesc     \tab specific conductance  data flag descriptions from the source program                                                                                                                                         \tab NULL                       \cr
spcond_censorcode                   \tab the LAGOSUS assigned censor code for specific conductance that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself \tab NULL                       \cr
spcond_detectionlimit_legacyvalue   \tab specific conductance detection limit used by the source program                                                                                                                                              \tab NULL                       \cr
spcond_detectionlimit_legacyunit    \tab specific conductance detection limit unit used by the source program                                                                                                                                         \tab NULL                       \cr
spcond_method                       \tab specific conductance analytical procedure, from a standards body if available                                                                                                                                \tab NULL                       \cr
do_mgl                              \tab oxygen, dissolved, milligrams per liter                                                                                                                                                                      \tab Milligrams per Liter       \cr
do_qualifier_legacyflag             \tab oxygen, dissolved data flags from the source program                                                                                                                                                         \tab NULL                       \cr
do_qualifier_legacyflagdesc         \tab oxygen, dissolved data flag descriptions from the source program                                                                                                                                             \tab NULL                       \cr
do_censorcode                       \tab the LAGOSUS assigned censor code for dissolved oxygen that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself     \tab NULL                       \cr
do_detectionlimit_legacyvalue       \tab oxygen, dissolved detection limit used by the source program                                                                                                                                                 \tab NULL                       \cr
do_detectionlimit_legacyunit        \tab oxygen, dissolved detection limit unit used by the source program                                                                                                                                            \tab NULL                       \cr
do_method                           \tab oxygen, dissolved analytical procedure, from a standards body if available                                                                                                                                   \tab NULL                       \cr
mg_mgl                              \tab magnesium, dissolved, milligrams per liter                                                                                                                                                                   \tab Milligrams per Liter       \cr
mg_qualifier_legacyflag             \tab magnesium, dissolved data flags from the source program                                                                                                                                                      \tab NULL                       \cr
mg_qualifier_legacyflagdesc         \tab magnesium, dissolved data flag descriptions from the source program                                                                                                                                          \tab NULL                       \cr
mg_censorcode                       \tab the LAGOSUS assigned censor code for magnesium, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself \tab NULL                       \cr
mg_detectionlimit_legacyvalue       \tab magnesium, dissolved detection limit used by the source program                                                                                                                                              \tab NULL                       \cr
mg_detectionlimit_legacyunit        \tab magnesium, dissolved detection limit unit used by the source program                                                                                                                                         \tab NULL                       \cr
mg_method                           \tab magnesium, dissolved analytical procedure, from a standards body if available                                                                                                                                \tab NULL                       \cr
ph_eq                               \tab pH, equilibrated                                                                                                                                                                                             \tab NULL                       \cr
ph_eq_qualifier_legacyflag          \tab pH, equilibrated data flags from the source program                                                                                                                                                          \tab NULL                       \cr
ph_eq_qualifier_legacyflagdesc      \tab pH, equilibrated data flag descriptions from the source program                                                                                                                                              \tab NULL                       \cr
ph_eq_censorcode                    \tab the LAGOSUS assigned censor code for ph_eq that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself                \tab NULL                       \cr
ph_eq_detectionlimit_legacyvalue    \tab pH, equilibrated detection limit used by the source program                                                                                                                                                  \tab NULL                       \cr
ph_eq_detectionlimit_legacyunit     \tab pH, equilibrated detection limit unit used by the source program                                                                                                                                             \tab NULL                       \cr
ph_eq_method                        \tab pH, equilibrated analytical procedure, from a standards body if available                                                                                                                                    \tab NULL                       \cr
ph_field                            \tab pH, field or closed                                                                                                                                                                                          \tab NULL                       \cr
ph_field_qualifier_legacyflag       \tab pH, field or closed data flags from the source program                                                                                                                                                       \tab NULL                       \cr
ph_field_qualifier_legacyflagdesc   \tab pH, field or closed data flag descriptions from the source program                                                                                                                                           \tab NULL                       \cr
ph_field_censorcode                 \tab the LAGOSUS assigned censor code for ph_field that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself             \tab NULL                       \cr
ph_field_detectionlimit_legacyvalue \tab pH, field or closed detection limit used by the source program                                                                                                                                               \tab NULL                       \cr
ph_field_detectionlimit_legacyunit  \tab pH, field or closed detection limit unit used by the source program                                                                                                                                          \tab NULL                       \cr
ph_field_method                     \tab pH, field or closed analytical procedure, from a standards body if available                                                                                                                                 \tab NULL                       \cr
salinity_mgl                        \tab salinity, milligrams per liter                                                                                                                                                                               \tab Milligrams per Liter       \cr
salinity_qualifier_legacyflag       \tab salinity data flags from the source program                                                                                                                                                                  \tab NULL                       \cr
salinity_qualifier_legacyflagdesc   \tab salinity data flag descriptions from the source program                                                                                                                                                      \tab NULL                       \cr
salinity_censorcode                 \tab the LAGOSUS assigned censor code for salinity that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself             \tab NULL                       \cr
salinity_detectionlimit_legacyvalue \tab salinity detection limit used by the source program                                                                                                                                                          \tab NULL                       \cr
salinity_detectionlimit_legacyunit  \tab salinity detection limit unit used by the source program                                                                                                                                                     \tab NULL                       \cr
salinity_method                     \tab salinity analytical procedure, from a standards body if available                                                                                                                                            \tab NULL                       \cr
so4_mgl                             \tab sulfate, dissolved, milligrams per liter as SO4                                                                                                                                                              \tab Milligrams per Liter       \cr
so4_qualifier_legacyflag            \tab sulfate, dissolved data flags from the source program                                                                                                                                                        \tab NULL                       \cr
so4_qualifier_legacyflagdesc        \tab sulfate, dissolved data flag descriptions from the source program                                                                                                                                            \tab NULL                       \cr
so4_censorcode                      \tab the LAGOSUS assigned censor code for sulfate, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself   \tab NULL                       \cr
so4_detectionlimit_legacyvalue      \tab sulfate, dissolved detection limit used by the source program                                                                                                                                                \tab NULL                       \cr
so4_detectionlimit_legacyunit       \tab sulfate, dissolved detection limit unit used by the source program                                                                                                                                           \tab NULL                       \cr
so4_method                          \tab sulfate, dissolved analytical procedure, from a standards body if available                                                                                                                                  \tab NULL                       \cr
temp_degc                           \tab temperature, degrees Celsius                                                                                                                                                                                 \tab Degrees Celcius            \cr
temp_qualifier_legacyflag           \tab temperature data flags from the source program                                                                                                                                                               \tab NULL                       \cr
temp_qualifier_legacyflagdesc       \tab temperature data flag descriptions from the source program                                                                                                                                                   \tab NULL                       \cr
temp_censorcode                     \tab the LAGOSUS assigned censor code for temperature that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself          \tab NULL                       \cr
temp_detectionlimit_legacyvalue     \tab temperature detection limit used by the source program                                                                                                                                                       \tab NULL                       \cr
temp_detectionlimit_legacyunit      \tab temperature detection limit unit used by the source program                                                                                                                                                  \tab NULL                       \cr
temp_method                         \tab temperature analytical procedure, from a standards body if available                                                                                                                                         \tab NULL
}
}
\keyword{datasets}
