% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{site_contaminants}
\alias{site_contaminants}
\title{LAGOSUS Contaminant data}
\description{
LAGOSUS Contaminant data

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
eventid_epi                           \tab unique event identifier that is a combination of programid, lakeid, and date for each sampling event in LAGOS                                                                                                \tab NULL                                    \cr
site_legacyid                         \tab unique sample site identifier provided by the source of the data; this is not standardardized and each source dataset has its own system                                                                     \tab NULL                                    \cr
program_id                            \tab unique integer program identifier assigned by LAGOSUS                                                                                                                                                        \tab NULL                                    \cr
program_name                          \tab name of sampling program that collected the data                                                                                                                                                             \tab NULL                                    \cr
event_date                            \tab date, as YYYY-MM-DD, when the lake was sampled by the source program                                                                                                                                         \tab YYYY-MM-DD                              \cr
site_id                               \tab unique sample site identifier created by the LAGOSUS                                                                                                                                                         \tab NULL                                    \cr
lagoslakeid                           \tab unique lake identifier developed by LAGOSUS                                                                                                                                                                  \tab NULL                                    \cr
site_lat_decdeg                       \tab latitude of the sample site                                                                                                                                                                                  \tab Decimal Degrees                         \cr
site_lon_decdeg                       \tab longitude of the sample site                                                                                                                                                                                 \tab Decimal Degrees                         \cr
program_type                          \tab type of program that provided the data                                                                                                                                                                       \tab NULL                                    \cr
al_tot_ugl                            \tab aluminum, total, micrograms per liter                                                                                                                                                                        \tab Micrograms per Liter                    \cr
al_tot_qualifier_legacyflag           \tab aluminum, total data flags from the source program                                                                                                                                                           \tab NULL                                    \cr
al_tot_qualifier_legacyflagdesc       \tab aluminum, total data flag descriptions from the source program                                                                                                                                               \tab NULL                                    \cr
al_tot_censorcode                     \tab the LAGOSUS assigned censor code for aluminum, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself      \tab NULL                                    \cr
al_tot_detectionlimit_legacyvalue     \tab aluminum, total detection limit used by the source program                                                                                                                                                   \tab NULL                                    \cr
al_tot_detectionlimit_legacyunit      \tab aluminum, total detection limit unit used by the source program                                                                                                                                              \tab NULL                                    \cr
al_tot_method                         \tab aluminum, total analytical procedure, from a standards body if available                                                                                                                                     \tab NULL                                    \cr
al_diss_ugl                           \tab aluminum, dissolved, micrograms per liter                                                                                                                                                                    \tab Micrograms per Liter                    \cr
al_diss_qualifier_legacyflag          \tab aluminum, dissolved data flags from the source program                                                                                                                                                       \tab NULL                                    \cr
al_diss_qualifier_legacyflagdesc      \tab aluminum, dissolved data flag descriptions from the source program                                                                                                                                           \tab NULL                                    \cr
al_diss_censorcode                    \tab the LAGOSUS assigned censor code for aluminum, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself  \tab NULL                                    \cr
al_diss_detectionlimit_legacyvalue    \tab aluminum, dissolved detection limit used by the source program                                                                                                                                               \tab NULL                                    \cr
al_diss_detectionlimit_legacyunit     \tab aluminum, dissolved detection limit unit used by the source program                                                                                                                                          \tab NULL                                    \cr
al_diss_method                        \tab aluminum, dissolved analytical procedure, from a standards body if available                                                                                                                                 \tab NULL                                    \cr
as_tot_ugl                            \tab arsenic, total, micrograms per liter                                                                                                                                                                         \tab Micrograms per Liter                    \cr
as_tot_qualifier_legacyflag           \tab arsenic, total data flags from the source program                                                                                                                                                            \tab NULL                                    \cr
as_tot_qualifier_legacyflagdesc       \tab arsenic, total data flag descriptions from the source program                                                                                                                                                \tab NULL                                    \cr
as_tot_censorcode                     \tab the LAGOSUS assigned censor code for arsenic, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself       \tab NULL                                    \cr
as_tot_detectionlimit_legacyvalue     \tab arsenic, total detection limit used by the source program                                                                                                                                                    \tab NULL                                    \cr
as_tot_detectionlimit_legacyunit      \tab arsenic, total detection limit unit used by the source program                                                                                                                                               \tab NULL                                    \cr
as_tot_method                         \tab arsenic, total analytical procedure, from a standards body if available                                                                                                                                      \tab NULL                                    \cr
as_diss_ugl                           \tab arsenic, dissolved, micrograms per liter                                                                                                                                                                     \tab Micrograms per Liter                    \cr
as_diss_qualifier_legacyflag          \tab arsenic, dissolved data flags from the source program                                                                                                                                                        \tab NULL                                    \cr
as_diss_qualifier_legacyflagdesc      \tab arsenic, dissolved data flag descriptions from the source program                                                                                                                                            \tab NULL                                    \cr
as_diss_censorcode                    \tab the LAGOSUS assigned censor code for arsenic, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself   \tab NULL                                    \cr
as_diss_detectionlimit_legacyvalue    \tab arsenic, dissolved detection limit used by the source program                                                                                                                                                \tab NULL                                    \cr
as_diss_detectionlimit_legacyunit     \tab arsenic, dissolved detection limit unit used by the source program                                                                                                                                           \tab NULL                                    \cr
as_diss_method                        \tab arsenic, dissolved analytical procedure, from a standards body if available                                                                                                                                  \tab NULL                                    \cr
atz_tot_ugl                           \tab atrazine, total, micrograms per liter                                                                                                                                                                        \tab Micrograms per Liter                    \cr
atz_tot_qualifier_legacyflag          \tab atrazine, total data flags from the source program                                                                                                                                                           \tab NULL                                    \cr
atz_tot_qualifier_legacyflagdesc      \tab atrazine, total data flag descriptions from the source program                                                                                                                                               \tab NULL                                    \cr
atz_tot_censorcode                    \tab the LAGOSUS assigned censor code for atrazine, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself      \tab NULL                                    \cr
atz_tot_detectionlimit_legacyvalue    \tab atrazine, total detection limit used by the source program                                                                                                                                                   \tab NULL                                    \cr
atz_tot_detectionlimit_legacyunit     \tab atrazine, total detection limit unit used by the source program                                                                                                                                              \tab NULL                                    \cr
atz_tot_method                        \tab atrazine, total analytical procedure, from a standards body if available                                                                                                                                     \tab NULL                                    \cr
atz_diss_ugl                          \tab atrazine, dissolved, micrograms per liter                                                                                                                                                                    \tab Micrograms per Liter                    \cr
atz_diss_qualifier_legacyflag         \tab atrazine, dissolved data flags from the source program                                                                                                                                                       \tab NULL                                    \cr
atz_diss_qualifier_legacyflagdesc     \tab atrazine, dissolved data flag descriptions from the source program                                                                                                                                           \tab NULL                                    \cr
atz_diss_censorcode                   \tab the LAGOSUS assigned censor code for atrazine, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself  \tab NULL                                    \cr
atz_diss_detectionlimit_legacyvalue   \tab atrazine, dissolved detection limit used by the source program                                                                                                                                               \tab NULL                                    \cr
atz_diss_detectionlimit_legacyunit    \tab atrazine, dissolved detection limit unit used by the source program                                                                                                                                          \tab NULL                                    \cr
atz_diss_method                       \tab atrazine, dissolved analytical procedure, from a standards body if available                                                                                                                                 \tab NULL                                    \cr
ecoli_cfu100ml                        \tab E. coli CFU, colony forming units per 100ml                                                                                                                                                                  \tab Colony Forming Units per 100 Milliliters\cr
ecoli_cfu_qualifier_legacyflag        \tab E. coli CFU data flags from the source program                                                                                                                                                               \tab NULL                                    \cr
ecoli_cfu_qualifier_legacyflagdesc    \tab E. coli CFU  data flag descriptions from the source program                                                                                                                                                  \tab NULL                                    \cr
ecoli_cfu_censorcode                  \tab the LAGOSUS assigned censor code for E. coli, cfu that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself         \tab NULL                                    \cr
ecoli_cfu_detectionlimit_legacyvalue  \tab E. coli CFU  detection limit used by the source program                                                                                                                                                      \tab NULL                                    \cr
ecoli_cfu_detectionlimit_legacyunit   \tab E. coli CFU  detection limit unit used by the source program                                                                                                                                                 \tab NULL                                    \cr
ecoli_cfu_method                      \tab E. coli CFU  analytical procedure, from a standards body if available                                                                                                                                        \tab NULL                                    \cr
ecoli_mpn100ml                        \tab E. coli MPN, most probable number per 100ml                                                                                                                                                                  \tab Most Probable Number per 100 Milliliters\cr
ecoli_mpn_qualifier_legacyflag        \tab E. coli MPN data flags from the source program                                                                                                                                                               \tab NULL                                    \cr
ecoli_mpn_qualifier_legacyflagdesc    \tab E. coli MPN data flag descriptions from the source program                                                                                                                                                   \tab NULL                                    \cr
ecoli_mpn_censorcode                  \tab the LAGOSUS assigned censor code for E. coli, MPN that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself         \tab NULL                                    \cr
ecoli_mpn_detectionlimit_legacyvalue  \tab E. coli MPN detection limit used by the source program                                                                                                                                                       \tab NULL                                    \cr
ecoli_mpn_detectionlimit_legacyunit   \tab E. coli MPN detection limit unit used by the source program                                                                                                                                                  \tab NULL                                    \cr
ecoli_mpn_method                      \tab E. coli MPN analytical procedure, from a standards body if available                                                                                                                                         \tab NULL                                    \cr
hg_tot_ngl                            \tab mercury, total, nanograms per liter                                                                                                                                                                          \tab Nanograms per Liter                     \cr
hg_tot_qualifier_legacyflag           \tab mercury, total data flags from the source program                                                                                                                                                            \tab NULL                                    \cr
hg_tot_qualifier_legacyflagdesc       \tab mercury, total data flag descriptions from the source program                                                                                                                                                \tab NULL                                    \cr
hg_tot_censorcode                     \tab the LAGOSUS assigned censor code for mercury, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself       \tab NULL                                    \cr
hg_tot_detectionlimit_legacyvalue     \tab mercury, total detection limit used by the source program                                                                                                                                                    \tab NULL                                    \cr
hg_tot_detectionlimit_legacyunit      \tab mercury, total detection limit unit used by the source program                                                                                                                                               \tab NULL                                    \cr
hg_tot_method                         \tab mercury, total analytical procedure, from a standards body if available                                                                                                                                      \tab NULL                                    \cr
hg_totdiss_ngl                        \tab mercury, total dissolved, nanograms per liter                                                                                                                                                                \tab Nanograms per Liter                     \cr
hg_totdiss_qualifier_legacyflag       \tab mercury, total dissolved data flags from the source program                                                                                                                                                  \tab NULL                                    \cr
hg_totdiss_qualifier_legacyflagdesc   \tab mercury, total dissolved data flag descriptions from the source program                                                                                                                                      \tab NULL                                    \cr
hg_totdiss_censorcode                 \tab the LAGOSUS assigned censor code for mercury, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself   \tab NULL                                    \cr
hg_totdiss_detectionlimit_legacyvalue \tab mercury, total dissolved detection limit used by the source program                                                                                                                                          \tab NULL                                    \cr
hg_totdiss_detectionlimit_legacyunit  \tab mercury, total dissolved detection limit unit used by the source program                                                                                                                                     \tab NULL                                    \cr
hg_totdiss_method                     \tab mercury, total dissolved analytical procedure, from a standards body if available                                                                                                                            \tab NULL                                    \cr
mehg_tot_ngl                          \tab methylmercury, total, nanograms per liter                                                                                                                                                                    \tab Nanograms per Liter                     \cr
mehg_tot_qualifier_legacyflag         \tab methylmercury, total data flags from the source program                                                                                                                                                      \tab NULL                                    \cr
mehg_tot_qualifier_legacyflagdesc     \tab methylmercury, total data flag descriptions from the source program                                                                                                                                          \tab NULL                                    \cr
mehg_tot_censorcode                   \tab the LAGOSUS assigned censor code for methylmercury, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself \tab NULL                                    \cr
mehg_tot_detectionlimit_legacyvalue   \tab methylmercury, total detection limit used by the source program                                                                                                                                              \tab NULL                                    \cr
mehg_tot_detectionlimit_legacyunit    \tab methylmercury, total detection limit unit used by the source program                                                                                                                                         \tab NULL                                    \cr
mehg_tot_method                       \tab methylmercury, total analytical procedure, from a standards body if available                                                                                                                                \tab NULL                                    \cr
se_tot_ugl                            \tab selenium, total, micrograms per liter                                                                                                                                                                        \tab Micrograms per Liter                    \cr
se_tot_qualifier_legacyflag           \tab selemium, total data flags from the source program                                                                                                                                                           \tab NULL                                    \cr
se_tot_qualifier_legacyflagdesc       \tab selemium, total data flag descriptions from the source program                                                                                                                                               \tab NULL                                    \cr
se_tot_censorcode                     \tab the LAGOSUS assigned censor code for seleneium, total that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself     \tab NULL                                    \cr
se_tot_detectionlimit_legacyvalue     \tab selenium, total detection limit used by the source program                                                                                                                                                   \tab NULL                                    \cr
se_tot_detectionlimit_legacyunit      \tab selenium, total detection limit unit used by the source program                                                                                                                                              \tab NULL                                    \cr
se_tot_method                         \tab selenium, total analytical procedure, from a standards body if available                                                                                                                                     \tab NULL                                    \cr
se_diss_ugl                           \tab selenium, dissolved, micrograms per liter                                                                                                                                                                    \tab Micrograms per Liter                    \cr
se_diss_qualifier_legacyflag          \tab selenium, dissolved data flags from the source program                                                                                                                                                       \tab NULL                                    \cr
se_diss_qualifier_legacyflagdesc      \tab selenium, dissolved data flag descriptions from the source program                                                                                                                                           \tab NULL                                    \cr
se_diss_censorcode                    \tab the LAGOSUS assigned censor code for selenium, dissolved that indicates whether the value is censored depending on information from the source program data flags, the detection limit, or the value itself  \tab NULL                                    \cr
se_diss_detectionlimit_legacyvalue    \tab selenium, dissolved detection limit used by the source program                                                                                                                                               \tab NULL                                    \cr
se_diss_detectionlimit_legacyunit     \tab selenium, dissolved detection limit unit used by the source program                                                                                                                                          \tab NULL                                    \cr
se_diss_method                        \tab selenium, dissolved analytical procedure, from a standards body if available                                                                                                                                 \tab NULL
}
}
\keyword{datasets}
