% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{site_information}
\alias{site_information}
\title{LAGOSUS LIMNO information}
\description{
LAGOSUS LIMNO information

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid     \tab unique lake identifier developed by LAGOSUS                                                                                              \tab NULL           \cr
site_id         \tab unique sample site identifier created by the LAGOSUS                                                                                     \tab NULL           \cr
site_legacyid   \tab unique sample site identifier provided by the source of the data; this is not standardardized and each source dataset has its own system \tab NULL           \cr
site_legacyname \tab sample site name provided by the source of the data                                                                                      \tab NULL           \cr
site_lat_decdeg \tab latitude of the sample site                                                                                                              \tab Decimal Degrees\cr
site_lon_decdeg \tab longitude of the sample site                                                                                                             \tab Decimal Degrees
}
}
\keyword{datasets}
