% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{source_table_limno}
\alias{source_table_limno}
\title{LAGOSUS LIMNO Data sources}
\description{
LAGOSUS LIMNO Data sources

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
program_id            \tab unique integer program identifier assigned by LAGOSUS                        \tab NULL      \cr
program_name          \tab name of sampling program that collected the data                             \tab NULL      \cr
program_description   \tab description of sampling program that collected the data                      \tab NULL      \cr
program_type          \tab type of program that provided the data                                       \tab NULL      \cr
program_startdate     \tab start date, as YYYY-MM-DD, of data provided by the source program            \tab YYYY-MM-DD\cr
program_enddate       \tab end date, as YYYY-MM-DD, of data provided by the source program              \tab YYYY-MM-DD\cr
program_analytelist   \tab list of analytes collected by the source program                             \tab NULL      \cr
program_lake_n        \tab number of lakes sampled by the source program                                \tab Count     \cr
program_wqpaccess     \tab data for the program was downloaded from the USGS Water Quality Portal (WQP) \tab NULL      \cr
program_lagosplatform \tab data provided by the program is included in LAGOSUS or LAGOSNE               \tab NULL
}
}
\keyword{datasets}
