% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{zone_soils}
\alias{zone_soils}
\title{LAGOSUS Soils data}
\description{
LAGOSUS Soils data

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lith_alkalinevolcanic_pct     \tab percent of zone with lithology described as alkaline intrusive volcanic rock                                       \tab pct   \cr
lith_alluviumcoastalfine_pct  \tab percent of zone with lithology described as alluvium and fine textured coastal zone sediment                       \tab pct   \cr
lith_carbonateresid_pct       \tab percent of zone with lithology described as carbonate residual material                                            \tab pct   \cr
lith_coastalzonecoarse_pct    \tab percent of zone with lithology described as coastal zone sediment, coarse textured                                 \tab pct   \cr
lith_colluvial_pct            \tab percent of zone with lithology described as colluvial sediment                                                     \tab pct   \cr
lith_eoliancoarse_pct         \tab percent of zone with lithology described as eolian sediment, coarse textured; for example sand dunes               \tab pct   \cr
lith_eolianfine_pct           \tab percent of zone with lithology described as eolian sediment, fine textured; for example glacial loess              \tab pct   \cr
lith_extrusivevolcanic_pct    \tab percent of zone with lithology described as extrusive volcanic rock                                                \tab pct   \cr
lith_glaciallakefine_pct      \tab percent of zone with lithology described as glacial lake sediment, fine textured                                   \tab pct   \cr
lith_glacialoutwashcoarse_pct \tab percent of zone with lithology described as glacial outwash and glacial lake sediment, coarse textured             \tab pct   \cr
lith_glacialtillclayey_pct    \tab percent of zone with lithology described as glacial till, clayey                                                   \tab pct   \cr
lith_glacialtillcoarse_pct    \tab percent of zone with lithology described as glacial till, coarse textured                                          \tab pct   \cr
lith_glacialtillloamy_pct     \tab percent of zone with lithology described as glacial till, loamy                                                    \tab pct   \cr
lith_hydricpeatmuck_pct       \tab percent of zone with lithology described as hydric, peat and muck                                                  \tab pct   \cr
lith_noncarbonateresid_pct    \tab percent of zone with lithology described as noncarbonate residual material                                         \tab pct   \cr
lith_salinelake_pct           \tab percent of zone with lithology described as saline lake sediment                                                   \tab pct   \cr
lith_silicicresid_pct         \tab percent of zone with lithology described as silicic residual material                                              \tab pct   \cr
lith_water_pct                \tab percent of zone with lithology described as water                                                                  \tab pct   \cr
soil_clay_pct                 \tab average percentage mass fraction of clay, 0 to 2 micrometers, in the 0 to 5 cm depth soil layer within the zone    \tab pct   \cr
soil_coarse_pct               \tab average percentage by volume of coarse fragments in the 0 to 5 cm soil depth layer within the zone                 \tab pct   \cr
soil_depthtobedrock_cm        \tab average absolute depth to bedrock within the zone                                                                  \tab cm    \cr
soil_kffact                   \tab average soil erodibility factor, not adjusted for the effect of rock fragments, within the zone                    \tab NULL  \cr
soil_orgcarbon_gperkg         \tab average organic carbon content, fine earth fraction, in the  0 to 5 cm soil layer within the zone                  \tab gperkg\cr
soil_sand_pct                 \tab average percentage mass fraction of sand, 50 to 200 micrometers, in the 0 to 5 cm depth soil layer within the zone \tab pct   \cr
soil_silt_pct                 \tab average percentage mass fraction of silt, 2 to 50 micrometers, in the 0 to 5 cm depth soil layer within the zone   \tab pct
}
}
\keyword{datasets}
