% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{zone_terrain}
\alias{zone_terrain}
\title{LAGOSUS Terrain data}
\description{
LAGOSUS Terrain data

\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
glaciatedlatewisc_pct             \tab glaciation status during the Late Wisconsin glaciation                    \tab pct\cr
landform_breaksfoothills_pct      \tab percent of zone with landform described as breaks or foothills            \tab pct\cr
landform_drainagechannel_pct      \tab percent of zone with landform described as drainage channel               \tab pct\cr
landform_escarpment_pct           \tab percent of zone with landform described as escarpment                     \tab pct\cr
landform_flatplain_pct            \tab percent of zone with landform described as flat plain                     \tab pct\cr
landform_highmountainscanyons_pct \tab percent of zone with landform described as high mountains or deep canyons \tab pct\cr
landform_hills_pct                \tab percent of zone with landform described as hills                          \tab pct\cr
landform_irregularplain_pct       \tab percent of zone with landform described as irregular plain                \tab pct\cr
landform_lowhills_pct             \tab percent of zone with landform described as low hills                      \tab pct\cr
landform_lowmountains_pct         \tab percent of zone with landform described as low mountains                  \tab pct\cr
landform_smoothplain_pct          \tab percent of zone with landform described as smooth plain                   \tab pct
}
}
\keyword{datasets}
