% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{lagosus_get}
\alias{lagosus_get}
\title{Retrieve LAGOSUS flat files}
\usage{
lagosus_get(
  locus_version = NA,
  locus_folder = NA,
  locus_overwrite = FALSE,
  limno_version = NA,
  limno_folder = NA,
  limno_overwrite = FALSE,
  geo_version = NA,
  geo_folder = NA,
  geo_overwrite = FALSE,
  depth_version = NA,
  depth_folder = NA,
  depth_overwrite = FALSE,
  dest_folder = NA
)
}
\arguments{
\item{locus_version}{character LAGOSUS database version string}

\item{locus_folder}{file.path to locus export folder}

\item{locus_overwrite}{logical overwrite existing data?}

\item{limno_version}{character LAGOSUS database version string}

\item{limno_folder}{file.path to limno export folder}

\item{limno_overwrite}{logical overwrite existing data?}

\item{geo_version}{character LAGOSUS database version string}

\item{geo_folder}{file.path to geo export folder}

\item{geo_overwrite}{logical overwrite existing data?}

\item{depth_version}{character LAGOSUS database version string}

\item{depth_folder}{file.path to depth export folder}

\item{depth_overwrite}{logical overwrite existing data?}

\item{dest_folder}{file.path recommended to set as the path returned by
\code{\link[LAGOSUS]{lagosus_path}}}
}
\description{
Retrieve LAGOSUS flat files from EDI.
}
\examples{
\dontrun{
# default to latest version
lagosus_get(dest_folder = LAGOSUS::lagosus_path())

# get previous version(s)
# - recommended to install corresponding package version
# - See 'Legacy Versions' section of the README for instructions
}
}
