% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{lake_characteristics}
\alias{lake_characteristics}
\title{LAGOSUS Lake characteristics}
\description{
\tabular{ll}{
  lagoslakeid                     \tab LAGOS sourced unique identifier for the focal lake                                                                                                                                       \cr
  lake_elevation_m                \tab the elevation of the lake polygon centroid, in meters (referenced to the North American Vertical Datum of 1988 (NAVD88)) and obtained from the National Elevation Dataset                \cr
  lake_waterarea_ha               \tab surface area of lake waterbody polygon from NHD (excludes islands)                                                                                                                       \cr
  lake_totalarea_ha               \tab surface area within outer boundary of lake waterbody polygon from NHD (includes islands)                                                                                                 \cr
  lake_islandarea_ha              \tab surface area of islands within outer boundary of lake waterbody polygon from NHD                                                                                                         \cr
  lake_perimeter_m                \tab perimeter of outer boundary of lake waterbody polygon from the NHD (excludes islands)                                                                                                    \cr
  lake_islandperimeter_m          \tab perimeter of islands within the lake waterbody polygon from the NHD                                                                                                                      \cr
  lake_shorelinedevfactor         \tab shoreline development factor calculated as the lake_perimeter_m / (2*√(π*lake_waterarea_ha*10000))                                                                                       \cr
  lake_mbgconhull_length_m        \tab the longest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon                                                                   \cr
  lake_mbgconhull_width_m         \tab the shortest distance between any two vertices of the minimum bounding geometry convex hull enclosing the zonal polygon                                                                  \cr
  lake_mbgconhull_orientation_deg \tab orientation (from 0 to 180 degrees) of the line connecting the antipodal pairs defining the longest distance within the minimum bounding geometry convex hull enclosing the zonal polygon\cr
  lake_mbgrect_length_m           \tab length of the minimum bounding geometry rectangle by area enclosing the zonal polygon                                                                                                    \cr
  lake_mbgrect_width_m            \tab width of the minimum bounding geometry rectangle by area enclosing the zonal polygon                                                                                                     \cr
  lake_mbgrect_arearatio          \tab ratio between lake_waterarea_ha (* 10000 m^2/ha) and the area of the minimum bounding rectangle by area of the zonal polygon                                                             \cr
  lake_meanwidth_m                \tab lake water area (* 10000 m^2/ha) divided by the convex hull length of the focal lake polygon                                                                                             \cr
  lake_connectivity_class         \tab hydrologic connectivity class of the focal lake determined from the NHD network considering both permanent and intermittent/ephemeral flow                                               \cr
  lake_connectivity_fluctuates    \tab indicates whether the lake's connectivity classification depends on non-permanent flow                                                                                                   \cr
  lake_connectivity_permanent     \tab hydrologic connectivity class of the focal lake determined from the NHD network considering only permanent flow                                                                          \cr
  lake_lakes1ha_upstream_ha       \tab total area of lakes >= 1 ha upstream of the focal lake, connected via surface streams                                                                                                    \cr
  lake_lakes4ha_upstream_ha       \tab total area of lakes >= 4 ha upstream of the focal lake, connected via surface streams                                                                                                    \cr
  lake_lakes10ha_upstream_ha      \tab total area of lakes >= 10 ha upstream of the focal lake, connected via surface streams                                                                                                   \cr
  lake_lakes1ha_upstream_n        \tab count of lakes >= 1 ha upstream of the focal lake, connected via surface streams                                                                                                         \cr
  lake_lakes4ha_upstream_n        \tab count of lakes >= 4 ha upstream of the focal lake, connected via surface streams                                                                                                         \cr
  lake_lakes10ha_upstream_n       \tab number per ha of lakes >= 10 ha upstream of the focal lake, connected via surface streams                                                                                                \cr
  lake_glaciatedlatewisc          \tab glaciation status during the Late Wisconsin glaciation
}
}
\keyword{datasets}
