% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{lake_information}
\alias{lake_information}
\title{LAGOSUS Lake information}
\description{
Includes information about lake identifiers used in LAGOS-US, information
about lake location (i.e., lat, lon, and elevation), and the zone IDs for
each spatial division containing the lake (e.g., state, county, hydrologic
unit).
}
\details{
\tabular{lllll}{
\bold{Variable name} \tab \bold{Variable name group} \tab \bold{Lagosne variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid        \tab lagoslakeid   \tab lagoslakeid   \tab unique lake identifier developed by LAGOS-US                                                                                                                                                                                    \tab               \cr
lake_nhdid         \tab nhdid         \tab nhdid         \tab the unique Permanent_Identifier from the NHD for each LAGOS lake                                                                                                                                                                \tab               \cr
lake_nhdfcode      \tab nhdfcode      \tab nhd_fcode     \tab NHD five-digit integer code with feature code plus additional characteristics or values                                                                                                                                         \tab               \cr
lake_nhdftype      \tab nhdftype      \tab nhd_ftype     \tab NHD three-digit integer code providing a unique identifier of feature type for the waterbody                                                                                                                                    \tab               \cr
lake_reachcode     \tab reachcode     \tab               \tab reach code assigned to the water feature by the NHD                                                                                                                                                                             \tab               \cr
lake_namegnis      \tab namegnis      \tab gnis_name     \tab lake name from the GNIS database                                                                                                                                                                                                \tab               \cr
lake_namelagos     \tab namelagos     \tab lagosname1    \tab lake name from a combination of data sources; examples are GNIS, WQP, etc.                                                                                                                                                      \tab               \cr
lake_onlandborder  \tab onlandborder  \tab               \tab flag indicating lake polygon is adjacent to or crosses the border with Canada or Mexico                                                                                                                                         \tab               \cr
lake_ismultipart   \tab ismultipart   \tab               \tab flag indicating that the focal lake polygon is multipart                                                                                                                                                                        \tab               \cr
lake_missingws     \tab missingws     \tab               \tab flag indicating that the WS for the lake was not delineated                                                                                                                                                                     \tab               \cr
lake_shapeflag     \tab shapeflag     \tab               \tab flag indicating lake polygon shape is angular, either triangular or rectangular, or is elongate and very thin relative to length; respectively may indicate the lake is less likely to be natural in origin or is more riverine \tab               \cr
lake_lat_decdeg    \tab lat_decdeg    \tab nhd_lat       \tab latitude of central point inside the NHD lake polygon in decimal degrees; NAD83 projection                                                                                                                                      \tab decimal degree\cr
lake_lon_decdeg    \tab lon_decdeg    \tab nhd_long      \tab longitude of  central point inside the NHD lake polygon in decimal degrees; NAD83 projection                                                                                                                                    \tab decimal degree\cr
lake_elevation_m   \tab elevation_m   \tab elevation_m   \tab the elevation of the lake polygon central point, in meters; referenced to the North American Vertical Datum of 1988 or NAVD88 and obtained from the National Elevation Dataset                                                  \tab meter         \cr
lake_centroidstate \tab centroidstate \tab               \tab two-letter postal abbreviation of the state containing the lake central point                                                                                                                                                   \tab               \cr
lake_states        \tab states        \tab               \tab two-letter postal abbreviations of all states intersecting the focal lake polygon                                                                                                                                               \tab               \cr
lake_county        \tab county        \tab               \tab name of county containing the focal lake polygon centroid                                                                                                                                                                       \tab               \cr
lake_countyfips    \tab countyfips    \tab               \tab Federal Information Processing Standard, FIPS, code for the county containing the focal lake polygon centroid                                                                                                                   \tab               \cr
lake_huc12         \tab huc12         \tab               \tab the NHD code for the HUC12 containing the lake centroid                                                                                                                                                                         \tab               \cr
buff100_zoneid     \tab zoneid        \tab               \tab unique 100 m buffer identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                                                                                     \tab               \cr
buff500_zoneid     \tab zoneid        \tab               \tab unique 500 m buffer identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                                                                                     \tab               \cr
ws_zoneid          \tab zoneid        \tab iws_zoneid    \tab unique WS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                                                                                     \tab               \cr
nws_zoneid         \tab zoneid        \tab               \tab unique NWS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                                                                                    \tab               \cr
hu12_zoneid        \tab zoneid        \tab hu12_zoneid   \tab unique identifier assigned by LAGOS-US for zones in the spatial division HU12                                                                                                                                                   \tab               \cr
hu8_zoneid         \tab zoneid        \tab hu8_zoneid    \tab unique identifier assigned by LAGOS-US for zones in the spatial division HU8                                                                                                                                                    \tab               \cr
hu4_zoneid         \tab zoneid        \tab hu4_zoneid    \tab unique identifier assigned by LAGOS-US for zones in the spatial division HU4                                                                                                                                                    \tab               \cr
county_zoneid      \tab zoneid        \tab county_zoneid \tab unique identifier assigned by LAGOS-US for zones in the spatial division county                                                                                                                                                 \tab               \cr
state_zoneid       \tab zoneid        \tab state_zoneid  \tab unique identifier assigned by LAGOS-US for zones in the spatial division state                                                                                                                                                  \tab               \cr
epanutr_zoneid     \tab zoneid        \tab               \tab unique identifier assigned by LAGOS-US for zones in the spatial division EPA Nutrient Ecoregions                                                                                                                                \tab               \cr
omernik3_zoneid    \tab zoneid        \tab               \tab unique identifier assigned by LAGOS-US for zones in the spatial division Omernik Level III Ecoregions                                                                                                                           \tab               \cr
wwf_zoneid         \tab zoneid        \tab               \tab unique identifier assigned by LAGOS-US for zones in the spatial division World Wildlife Fund ecoregions                                                                                                                         \tab               \cr
mlra_zoneid        \tab zoneid        \tab               \tab unique identifier assigned by LAGOS-US for zones in the spatial division Major Land Areas                                                                                                                                       \tab               \cr
bailey_zoneid      \tab zoneid        \tab               \tab unique identifier assigned by LAGOS-US for zones in the spatial division Baileys Ecoregions                                                                                                                                     \tab               \cr
neon_zoneid        \tab zoneid        \tab               \tab unique identifier assigned by LAGOS-US for zones in the spatial division National Ecological Observation Network                                                                                                                \tab
}
}
\keyword{datasets}
