% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{lake_link}
\alias{lake_link}
\title{LAGOSUS Identifier links}
\description{
Includes single or multiple identifiers per lake from other commonly used
national-scale data products including: Water Quality Portal (WQP),
Geographic Names Information System (GNIS), NHD MR Plusv2, LAGOS-NE, and
the EPA National Lake Assessment surveys from 2007 and 2012.
}
\details{
\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                      \tab unique lake identifier developed by LAGOS-US                                                                                                                \tab NULL            \cr
lake_nhdid                       \tab the unique Permanent_Identifier from the NHD for each LAGOS lake                                                                                            \tab NULL            \cr
lake_reachcode                   \tab reach code assigned to the water feature by the NHD                                                                                                         \tab NULL            \cr
lake_namegnis                    \tab lake name from the GNIS database                                                                                                                            \tab NULL            \cr
lake_namelagos                   \tab lake name from a combination of data sources; examples are GNIS, WQP, etc.                                                                                  \tab NULL            \cr
lake_county                      \tab name of county containing the focal lake polygon centroid                                                                                                   \tab NULL            \cr
lake_countyfips                  \tab Federal Information Processing Standard, FIPS, code for the county containing the focal lake polygon centroid                                               \tab NULL            \cr
lake_lat_decdeg                  \tab latitude of centroid of the NHD lake polygon in decimal degrees; NAD83 projection                                                                           \tab decimal degree  \cr
lake_lon_decdeg                  \tab longitude of  centroid of the NHD lake polygon in decimal degrees; NAD83 projection                                                                         \tab decimal degree  \cr
lake_centroidstate               \tab two-letter postal abbreviation of the state containing the lake centroid                                                                                    \tab NULL            \cr
nhdhr_area_sqkm                  \tab area of areal feature based on Albers Equal Area, as assigned in NHD                                                                                        \tab square kilometer\cr
nhdhr_fdate                      \tab date of last feature modification in National Hydrography Dataset                                                                                           \tab NULL            \cr
nhdhr_gnisid                     \tab unique identifier assigned by GNIS                                                                                                                          \tab NULL            \cr
lagosus_legacysiteid             \tab unique site or lake identifier from the original LAGOS-US water quality source dataset--this is not standardized and each source dataset has its own system \tab NULL            \cr
lagosus_legacysitelabel          \tab column name or field name used by the source dataset to label the legacy lake identifier                                                                    \tab NULL            \cr
lagosus_legacyprogram            \tab program name assigned in LAGOS-US to identify the source dataset provider                                                                                   \tab NULL            \cr
wqp_monitoringlocationidentifier \tab a designator used to describe the unique name, number, or code assigned to identify the monitoring location                                                 \tab NULL            \cr
wqp_monitoringlocationname       \tab the designator specified by the sampling organization for the site at which sampling or other activities are conducted                                      \tab NULL            \cr
wqp_providername                 \tab source system that provided data to the Water Quality Portal; examples are NWIS, STORET, STEWARDS, etc.                                                     \tab NULL            \cr
nhdplusv2_comid                  \tab common identifier of the NHD Waterbody feature                                                                                                              \tab NULL            \cr
nhdplusv2_reachcode              \tab NHDplus v2 reach code assigned to the water feature.                                                                                                        \tab NULL            \cr
nhdplusv2_area_sqkm              \tab feature area in square kilometers, as assigned by NHD                                                                                                       \tab square kilometer\cr
lagosne_lagoslakeid              \tab unique lake identifier developed for LAGOS-NE                                                                                                               \tab NULL            \cr
lagosne_legacysiteid             \tab unique site/lake identifier from the original LAGOS-NE water quality source dataset--this is not standardized and each source dataset has its own system    \tab NULL            \cr
nla2007_siteid                   \tab lake site identifier from National Lakes Assessment 2007 survey                                                                                             \tab NULL            \cr
nla2012_siteid                   \tab lake site identifier from National Lakes Assessment 2012 survey                                                                                             \tab NULL            \cr
nhdplusv2_lakes_n                \tab count of NHDPlusV2 polygons matching the LAGOS-US lagoslakeid                                                                                               \tab count           \cr
lagosne_lakes_n                  \tab count of LAGOS-NE lagoslakeids matching the LAGOS-US lagoslakeid                                                                                            \tab count           \cr
wqp_sites_n                      \tab count of WQP sites within the LAGOS-US lake                                                                                                                 \tab count           \cr
lagosus_legacyids_n              \tab count of all LAGOS-US legacy ids linked to the lagoslakeid in this table; these ids include, for example, original sampling site identifiers                \tab count
}
}
\keyword{datasets}
