% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{lake_watersheds}
\alias{lake_watersheds}
\alias{watersheds}
\title{LAGOSUS Lake watersheds}
\description{
Includes identifiers, location, and geometry for the calculated watersheds (WS) and network
watersheds (NWS).
}
\details{
\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid                    \tab unique lake identifier developed by LAGOS-US                                                                                                                                                                                                                                                                     \tab NULL          \cr
ws_zoneid                      \tab unique WS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                                                                                                                                                                      \tab NULL          \cr
nws_zoneid                     \tab unique NWS watershed identifier assigned by LAGOS-US; identifier is equivalent to the focal lake lagoslakeid                                                                                                                                                                                                     \tab NULL          \cr
ws_subtype                     \tab watershed subtype class for the watershed; LC equals local catchment, DWS equals drainage-watershed, IDWS equals inter-drainage-lake watershed                                                                                                                                                                   \tab NULL          \cr
ws_equalsnws                   \tab flag indicating that the ws is equivalent to the nws                                                                                                                                                                                                                                                             \tab NULL          \cr
ws_onlandborder                \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                                                                                                                                                                                  \tab NULL          \cr
ws_oncoast                     \tab flag indicating the zone is adjacent to the coastline                                                                                                                                                                                                                                                            \tab NULL          \cr
ws_inusa_pct                   \tab percent of the zone polygon that is within the US border                                                                                                                                                                                                                                                         \tab percent       \cr
ws_includeshu4inlet            \tab flag indicating the WS extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                                                                                                                                                                            \tab NULL          \cr
ws_ismultipart                 \tab flag indicating that the focal zone polygon is multipart                                                                                                                                                                                                                                                         \tab NULL          \cr
ws_sliverflag                  \tab flag indicating extreme cases where the watershed is comprised of slivers due to flat or constrained terrain that may not accurately reflect surface drainage                                                                                                                                                    \tab NULL          \cr
ws_sinkerrorflag               \tab flag indicating an error that applies to a small subset of watersheds that erroneously include in the WS polygon an upstream lake with only in inlet that is greater than or equal to 10 ha, along with its respective WS                                                                                        \tab NULL          \cr
ws_divergenceerrorflag         \tab flag indicating an error that applies to a small subset of watersheds when the presence of a second lake greater than or equal to 10ha located below a common network divergence led to the erroneous erasure of an overlapping watershed area upstream of that divergence from the focal lake watershed         \tab NULL          \cr
nws_divergenceerrorflag        \tab flag indicating an error that applies to a small subset of watersheds when the presence of a second lake greater than or equal to 10ha located below a common network divergence led to the erroneous erasure of an overlapping watershed area upstream of that divergence from the focal lake network watershed \tab NULL          \cr
nws_onlandborder               \tab flag indicating zone is adjacent to or crosses the border with Canada or Mexico                                                                                                                                                                                                                                  \tab NULL          \cr
nws_oncoast                    \tab flag indicating the zone is adjacent to the coastline                                                                                                                                                                                                                                                            \tab NULL          \cr
nws_inusa_pct                  \tab percent of the zone polygon that is within the US border                                                                                                                                                                                                                                                         \tab percent       \cr
nws_includeshu4inlet           \tab flag indicating the NWS extends beyond the focal HU4 and is connected via an inlet to an adjoining HU4                                                                                                                                                                                                           \tab NULL          \cr
nws_ismultipart                \tab flag indicating that the focal zone polygon is multipart                                                                                                                                                                                                                                                         \tab NULL          \cr
ws_states                      \tab two-letter postal abbreviations of all states intersecting the zone polygon                                                                                                                                                                                                                                      \tab NULL          \cr
ws_focallakewaterarea_ha       \tab surface area of focal lake for which the WS is calculated                                                                                                                                                                                                                                                        \tab hectare       \cr
ws_area_ha                     \tab area of zone polygon                                                                                                                                                                                                                                                                                             \tab hectare       \cr
ws_perimeter_m                 \tab perimeter of zone polygon including perimeter of any internal holes                                                                                                                                                                                                                                              \tab meter         \cr
ws_lake_arearatio              \tab ratio between WS area and lake water area                                                                                                                                                                                                                                                                        \tab NULL          \cr
ws_mbgconhull_length_m         \tab the maximum possible distance between any two vertices of the minimum bounding geometry convex hull that encloses the WS polygon                                                                                                                                                                                 \tab meter         \cr
ws_mbgconhull_width_m          \tab the maximum width of the WS convex hull polygon along an axis perpendicular to the axis defined by the convex hull length                                                                                                                                                                                        \tab meter         \cr
ws_mbgconhull_orientation_deg  \tab orientation of the line defining the convex hull length; calculated as decimal degrees clockwise from north                                                                                                                                                                                                      \tab degree        \cr
ws_meanwidth_m                 \tab mean width of the WS polygon; calculated as ws_area_ha converted to hectares and divided by ws_mbgconhull_length_m                                                                                                                                                                                               \tab meter         \cr
ws_lat_decdeg                  \tab latitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection                                                                                                                                                                                                                          \tab decimaldegree \cr
ws_lon_decdeg                  \tab longitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection                                                                                                                                                                                                                         \tab decimaldegree \cr
nws_states                     \tab two-letter postal abbreviations of all states intersecting the zone polygon                                                                                                                                                                                                                                      \tab NULL          \cr
nws_focallakewaterarea_ha      \tab surface area of focal lake for which the NWS is calculated                                                                                                                                                                                                                                                       \tab hectare       \cr
nws_area_ha                    \tab area of zone polygon                                                                                                                                                                                                                                                                                             \tab hectare       \cr
nws_perimeter_m                \tab perimeter of zone polygon including perimeters of any internal holes                                                                                                                                                                                                                                             \tab meter         \cr
nws_lake_arearatio             \tab ratio between NWS area and lake water area                                                                                                                                                                                                                                                                       \tab NULL          \cr
nws_mbgconhull_length_m        \tab the maximum possible distance between any two vertices of the minimum bounding geometry convex hull that encloses the NWS polygon                                                                                                                                                                                \tab meter         \cr
nws_mbgconhull_width_m         \tab the maximum width of the NWS convex hull polygon along an axis perpendicular to the axis defined by the convex hull length                                                                                                                                                                                       \tab meter         \cr
nws_mbgconhull_orientation_deg \tab orientation of the line defining the convex hull length; calculated as decimal degrees clockwise from north                                                                                                                                                                                                      \tab degree        \cr
nws_meanwidth_m                \tab mean width of the NWS polygon; calculated as nws_area_ha converted to hectares and divided by nws_mbgconhull_length_m                                                                                                                                                                                            \tab meter         \cr
nws_lat_decdeg                 \tab latitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection                                                                                                                                                                                                                          \tab decimal degree\cr
nws_lon_decdeg                 \tab longitude of the centroid of the focal lake polygon in decimal degrees; NAD83 projection                                                                                                                                                                                                                         \tab decimal degree
}
}
\keyword{datasets}
