% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAGOSUS-package.R
\docType{data}
\name{locus_information}
\alias{locus_information}
\title{LAGOSUS Lake information}
\description{
Includes information about lake identifiers used in LAGOS-US, information
about lake location (i.e., lat, lon, and elevation), and the zone IDs for
each spatial division containing the lake (e.g., state, county, hydrologic
unit).
}
\details{
\tabular{lll}{
\bold{Variable name} \tab \bold{Variable description} \tab \bold{Units} \cr
lagoslakeid        \tab unique lake identifier developed by LAGOS-US                                                                                                                                                                              \tab               \cr
  lake_nhdid         \tab the unique 'Permanent_identifier' from the NHD for each LAGOS lake                                                                                                                                                        \tab               \cr
  lake_nhdfcode      \tab NHD five-digit integer code with feature code plus additional characteristics/values                                                                                                                                      \tab               \cr
  lake_nhdftype      \tab NHD three-digit integer code providing a unique identifier of feature type for the waterbody                                                                                                                              \tab               \cr
  lake_reachcode     \tab reach code assigned to the water feature by the NHD                                                                                                                                                                       \tab               \cr
  lake_namegnis      \tab lake name from the gnis database                                                                                                                                                                                          \tab               \cr
  lake_namelagos     \tab lake name from a combination of data sources (gnis, wqp, etc.)                                                                                                                                                            \tab               \cr
  lake_onlandborder  \tab flag indicating lake polygon is adjacent to or crosses the border with Canada or Mexico                                                                                                                                   \tab               \cr
  lake_ismultipart   \tab flag indicating that the focal lake polygon is multipart                                                                                                                                                                  \tab               \cr
  lake_missingws     \tab flag indicating that the lake's watershed (both ws and nws) was not delineated                                                                                                                                            \tab               \cr
  lake_shapeflag     \tab flag indicating lake polygon shape is excessively angular (e.g., triangle, rectangle) or elongate (very thin relative to length); may indicate the lake is not natural in origin (angular) or is more riverine (elongate) \tab               \cr
  lake_lat_decdeg    \tab latitude of centroid of the NHD lake polygon in decimal degrees (NAD83)                                                                                                                                                   \tab decimal degree\cr
  lake_lon_decdeg    \tab longitude of  centroid of the NHD lake polygon in decimal degrees (NAD83)                                                                                                                                                 \tab decimal degree\cr
  lake_elevation_m   \tab the elevation of the lake polygon centroid, in meters (referenced to the North American Vertical Datum of 1988 (NAVD88) and obtained from the National Elevation Dataset)                                                 \tab meter         \cr
  lake_centroidstate \tab abbreviation of state containing the lake centroid                                                                                                                                                                        \tab               \cr
  lake_states        \tab abbreviation(s) of state(s) intersecting the focal lake polygon                                                                                                                                                           \tab               \cr
  lake_county        \tab NULLme of county containing the focal lake polygon centroid                                                                                                                                                               \tab               \cr
  lake_countyfips    \tab Federal Information Processing Standard (FIPS) code for the county containing the focal lake polygon centroid                                                                                                             \tab               \cr
  lake_huc12         \tab the code for the HUC12 containing the lake centroid                                                                                                                                                                       \tab               \cr
  buff100_zoneid     \tab unique 100 m buffer identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid                                                                                                                             \tab               \cr
  buff500_zoneid     \tab unique 500 m buffer identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid                                                                                                                             \tab               \cr
  ws_zoneid          \tab unique watershed identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid                                                                                                                                \tab               \cr
  nws_zoneid         \tab unique network watershed identifier assigned by LAGOS-US, equivalent to the focal lake lagoslakeid                                                                                                                        \tab               \cr
  hu12_zoneid        \tab unique identifier assigned by LAGOS-US for zones in the spatial division HU12                                                                                                                                             \tab               \cr
  hu8_zoneid         \tab unique identifier assigned by LAGOS-US for zones in the spatial division HU8                                                                                                                                              \tab               \cr
  hu4_zoneid         \tab unique identifier assigned by LAGOS-US for zones in the spatial division HU4                                                                                                                                              \tab               \cr
  county_zoneid      \tab unique identifier assigned by LAGOS-US for zones in the spatial division county                                                                                                                                           \tab               \cr
  state_zoneid       \tab unique identifier assigned by LAGOS-US for zones in the spatial division state                                                                                                                                            \tab               \cr
  epanutr_zoneid     \tab unique identifier assigned by LAGOS-US for zones in the spatial division EPA Nutrient Ecoregions                                                                                                                          \tab               \cr
  omernik3_zoneid    \tab unique identifier assigned by LAGOS-US for zones in the spatial division Omernik Level III Ecoregions                                                                                                                     \tab               \cr
  wwf_zoneid         \tab unique identifier assigned by LAGOS-US for zones in the spatial division World Wildlife Fund ecoregions                                                                                                                   \tab               \cr
  mlra_zoneid        \tab unique identifier assigned by LAGOS-US for zones in the spatial division Major Land Areas                                                                                                                                 \tab               \cr
  bailey_zoneid      \tab unique identifier assigned by LAGOS-US for zones in the spatial division Bailey’s Ecoregions                                                                                                                              \tab               \cr
  neon_zoneid        \tab unique identifier assigned by LAGOS-US for zones in the spatial division National Ecological Observation Network                                                                                                          \tab               
}
}
\keyword{datasets}
